% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{add_pars_to_grids}
\alias{add_pars_to_grids}
\title{Add single-valued parameters to the pargrids}
\usage{
add_pars_to_grids(pars, gridlist, FLAGoverwrite = FALSE)
}
\arguments{
\item{pars}{character or numeric}

\item{gridlist}{list(fix.grid, est.grid)}

\item{FLAGoverwrite}{Overwrite existing parameters in fix.grid or est.grid?}
}
\value{
gridlist
}
\description{
Add single-valued parameters to the pargrids
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
