% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{approxExtrap}
\alias{approxExtrap}
\title{Extrapolation extension to approx}
\usage{
approxExtrap(x, y, xout, method = "linear", n = 50, rule = 2, f = 0,
  ties = "ordered", na.rm = FALSE)
}
\description{
from Hmisc package, built on approx
}
