% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataClass.R, R/prdClass.R
\name{as.data.frame.datalist}
\alias{as.data.frame.datalist}
\alias{as.data.frame.prdlist}
\title{Coerce to a Data Frame}
\usage{
\method{as.data.frame}{datalist}(x, ...)

\method{as.data.frame}{prdlist}(x, data = NULL, errfn = NULL, ...)
}
\arguments{
\item{x}{any R object}

\item{...}{additional arguments to be passed to or from methods}

\item{data}{data list oject}

\item{errfn}{obsfn object, the error model function to predict sigma}
}
\value{
a data frame
}
\description{
Coerce to a Data Frame
}
