% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqnClass.R
\name{as.data.frame.eqnlist}
\alias{as.data.frame.eqnlist}
\title{Coerce equation list into a data frame}
\usage{
\method{as.data.frame}{eqnlist}(x, ...)
}
\arguments{
\item{x}{object of class \link{eqnlist}}

\item{...}{other arguments}
}
\value{
a \code{data.frame} with columns "Description" (character), 
"Rate" (character), and one column per ODE state with the state names. 
The state columns correspond to the stoichiometric matrix.
}
\description{
Coerce equation list into a data frame
}
