% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqnClass.R
\name{as.eqnvec}
\alias{as.eqnvec}
\alias{as.eqnvec.character}
\alias{as.eqnvec.eqnlist}
\title{Coerce to an equation vector}
\usage{
as.eqnvec(x, ...)

\method{as.eqnvec}{character}(x = NULL, names = NULL, ...)

\method{as.eqnvec}{eqnlist}(x, ...)
}
\arguments{
\item{x}{object of class \code{character} or \code{eqnlist}}

\item{...}{arguments going to the corresponding methods}

\item{names}{character, the left-hand sides of the equation}
}
\value{
object of class \link{eqnvec}.
}
\description{
An equation list stores an ODE in a list format. The function
translates this list into the right-hand sides of the ODE.
}
\details{
If \code{x} is of class \code{eqnlist}, \link{getFluxes} is called and coerced
into a vector of equations.
}
