% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventClass.R
\name{as.eventlist}
\alias{as.eventlist}
\alias{as.eventlist.list}
\alias{as.eventlist.data.frame}
\title{Coerce to eventlist}
\usage{
as.eventlist(x, ...)

\method{as.eventlist}{list}(x, ...)

\method{as.eventlist}{data.frame}(x, ...)
}
\arguments{
\item{x}{list, data.frame}

\item{...}{not used}
}
\description{
Coerce to eventlist
}
