% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{as.obj}
\alias{as.obj}
\title{Generate dummy list of class \code{obj} from named numeric}
\usage{
as.obj(p)
}
\arguments{
\item{p}{Names numeric vector}
}
\value{
list with entries value (\code{0}),
gradient (\code{rep(0, length(p))}) and
hessian (\code{matrix(0, length(p), length(p))}) of class \code{obj}.
}
\description{
Generate dummy list of class \code{obj} from named numeric
}
\examples{
p <- c(A = 1, B = 2)
as.obj(p)
}

