% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objClass.R
\name{as.objlist}
\alias{as.objlist}
\title{Generate objective list from numeric vector}
\usage{
as.objlist(p)
}
\arguments{
\item{p}{Named numeric vector}
}
\value{
list with entries value (\code{0}), 
gradient (\code{rep(0, length(p))}) and 
hessian (\code{matrix(0, length(p), length(p))}) of class \code{obj}.
}
\description{
Generate objective list from numeric vector
}
\examples{
p <- c(A = 1, B = 2)
as.objlist(p)
}

