% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsWolfgang.R
\name{as.parvec.parframe}
\alias{as.parvec.parframe}
\title{Select a parameter vector from a parameter frame.}
\usage{
\method{as.parvec}{parframe}(x, index = 1, ...)
}
\arguments{
\item{x}{A parameter frame, e.g., the output of
\code{\link{as.parframe}}.}

\item{index}{Integer, the parameter vector with the \code{index}-th lowest
objective value.}

\item{...}{not used right now}
}
\value{
The parameter vector with the \option{index}-th lowest objective
  value.
}
\description{
Obtain a parameter vector from a parameter frame.
}
\details{
With this command, additional information included in the parameter
  frame as the objective value and the convergence state are removed and a
  parameter vector is returned. This parameter vector can be used to e.g.,
  evaluate an objective function.
  
  On selection, the parameters in the parameter frame are ordered such, that
  the parameter vector with the lowest objective value is at \option{index}
  1. Thus, the parameter vector with the \option{index}-th lowest objective
  value is easily obtained.
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}

