% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRtools.R
\name{as_IQRsysProjectMulti}
\alias{as_IQRsysProjectMulti}
\title{Creating an IQRsysProjectMulti project}
\usage{
as_IQRsysProjectMulti(input, FLAGrecursive = FALSE)
}
\arguments{
\item{input}{a folder where to find the projects are distributed in subfolders}
}
\value{
List of folders where projects were found. The output gets a class 
\code{IQRsysProjectMulti}.
}
\description{
Creating an IQRsysProjectMulti project
}
\seealso{
\code{\link{summary.IQRsysProjectMulti}}
}
