% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{attrs}
\alias{attrs}
\title{Select attributes.}
\usage{
attrs(x, atr = NULL, keep = TRUE)
}
\arguments{
\item{x}{The object to work on}

\item{atr}{An optional list of attributes which are either kept or removed. 
This parameter defaults to dim, dimnames, names,  col.names, and row.names.}

\item{keep}{For keep = TRUE, atr is a positive list on attributes which are 
kept, for keep = FALSE, \option{atr} are removed.}
}
\value{
x with selected attributes.
}
\description{
Select or discard attributes from an object.
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}

Mirjam Fehling-Kaschek, \email{mirjam.fehling@physik.uni-freiburg.de}
}
