% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BAdata}
\alias{BAdata}
\title{Time-course data for the Bile Acid model
 
The dataset used in the dMod publication: Kaschek D, Mader W, Fehling-Kaschek M, Rosenblatt M, 
Timmer J. Dynamic modeling, parameter estimation, and uncertainty analysis in R. 
Journal of Statistical Software. 2019 Apr 30;88:1-32.}
\description{
Time-course data for the Bile Acid model
 
The dataset used in the dMod publication: Kaschek D, Mader W, Fehling-Kaschek M, Rosenblatt M, 
Timmer J. Dynamic modeling, parameter estimation, and uncertainty analysis in R. 
Journal of Statistical Software. 2019 Apr 30;88:1-32.
}
\keyword{data}
