% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{blockdiagSymb}
\alias{blockdiagSymb}
\title{Embed two matrices into one blockdiagonal matrix}
\usage{
blockdiagSymb(M, N)
}
\arguments{
\item{M}{matrix of type character}

\item{N}{matrix of type character}
}
\value{
Matrix of type character containing M and N as upper left and lower right block
}
\description{
Embed two matrices into one blockdiagonal matrix
}

