% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{build_est.grid}
\alias{build_est.grid}
\title{Build the est.grid from prd, fixed.grid, conditional and condition.grid}
\usage{
build_est.grid(prd0, fixed.grid, conditional, condition.grid, iiv = NULL)
}
\arguments{
\item{prd0}{prdfn for 1 condition. Extracts parameters via getParameters}

\item{fixed.grid}{As specified in template_project_sysfit.R: data.frame(parname, partask, ids...)}

\item{conditional}{As specified in template_project_sysfit.R: data.frame(parname, covname, covvalue)}

\item{condition.grid}{from datalist}

\item{iiv}{parameters with iiv}
}
\value{
est.grid data.frame(parname, ids...)
}
\description{
Performs indiviudalization and localization of parameters
}
