% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R2CdeSolve.R
\name{bvptwpC.einspline}
\alias{bvptwpC.einspline}
\title{Interface to bvptwp()}
\usage{
\method{bvptwpC}{einspline}(yini = NULL, x, func, yend = NULL, parms,
  xguess = NULL, yguess = NULL, ...)
}
\arguments{
\item{yini}{named vector of type numeric. Initial values to be overwritten.}

\item{x}{vector of type numeric. Integration times}

\item{func}{return value from funC() with a boundary argument.}

\item{yend}{named vector of type numeric. End values to be overwritten.}

\item{parms}{named vector of type numeric. The dynamic parameters.}

\item{xguess}{vector of type numeric, the x values}

\item{yguess}{matrix with as many rows as variables and columns as x values}

\item{...}{further arguments going to \code{bvptwp()}}
}
\value{
matrix with times and states
}
\description{
Interface to bvptwp()
}
\details{
See bvpSolve-package for a full description of possible arguments
}

