% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{check_and_sanitize_prediction}
\alias{check_and_sanitize_prediction}
\title{Run some checks on the prediction in normL2_indiv}
\usage{
check_and_sanitize_prediction(prediction, data, cn, FLAGNaNInfwarnings)
}
\arguments{
\item{prediction}{prediction for condition cn}

\item{data}{datalist}

\item{cn}{condition name for which the prediction was made}

\item{FLAGNaNInfwarnings}{print warnings?}
}
\value{
the prediction with harmless NA's replaced by 0
}
\description{
If prediction is NA for observables which are not observed in a condition, they don't matter.
In this case, replace NA by 0, such that the error model can be evaluated.
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
