% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModFrameClass.R
\name{checkout_hypothesis}
\alias{checkout_hypothesis}
\title{Load one row of a dMod.frame into the .GlobalEnv}
\usage{
checkout_hypothesis(dMod.frame, hypothesis, prefix = "", suffix = "")
}
\arguments{
\item{dMod.frame}{A dMod.frame}

\item{hypothesis}{character or numeric. specifying the name  or the index of the hypothesis}

\item{prefix}{Prefix appended to the object names in .GlobalEnv}

\item{suffix}{Suffix appended to the object names in .GlobalEnv}
}
\description{
Load one row of a dMod.frame into the .GlobalEnv
}
\examples{
testframe <- dplyr::tibble(hypothesis = c("linear", "quadratic"),
                    plots = list(plot(1:10,1:10), plot(1:10,(1:10)^2)),
                    myfun = list(function(x,a) {a * x}, function(x,a) {a * x^2}),
                    a = c(1:2))


checkout_hypothesis(testframe, "quadratic", prefix = "quad")
quadplots
quadmyfun(1:10, quada)
}
