% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{classic2Indiv_trafo}
\alias{classic2Indiv_trafo}
\title{Title}
\usage{
classic2Indiv_trafo(trafo, cg)
}
\arguments{
\item{trafo}{base trafo}

\item{cg}{output of [indiv2Classic_gridlist2cond.grid()]}
}
\value{

}
\description{
Title
}
