% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateEquations.R
\name{combine}
\alias{combine}
\title{Combine several data.frames from csv read-in}
\usage{
combine(...)
}
\arguments{
\item{...}{data.frames from csv read-in}
}
\value{
A \code{data.frame} with columns "Description", "Rate" and one column for each
ODE state in the model. See \link{generateEquations}.
}
\description{
Combine several data.frames from csv read-in
}
\details{
This function is useful when separating the model into independent csv model files,
e.g.~a receptor model and several downstream paths. Then, the models can be recombined
into one model by \code{combine()}.
}

