% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trustL1.R
\name{constraintL1}
\alias{constraintL1}
\title{Soft L1 constraint on parameters}
\usage{
constraintL1(p, mu, lambda = 1, fixed = NULL)
}
\arguments{
\item{p}{Namec numeric, the parameter value}

\item{mu}{Named numeric, the prior values}

\item{lambda}{Named numeric of length of mu or numeric of length one.}

\item{fixed}{Named numeric with fixed parameter values (contribute to the prior value
but not to gradient and Hessian)}
}
\value{
List of class \code{obj}, i.e. objective value, gradient and Hessian as list.
}
\description{
Soft L1 constraint on parameters
}
\details{
Computes the constraint value 
\deqn{\lambda\|p-\mu\|}{lambda*abs(p-mu)}
and its derivatives with respect to p.
}
\examples{
p <- c(A = 1, B = 2, C = 3)
mu <- c(A = 0, B = 0)
lambda <- c(A = 0.1, B = 1)
constraintL1(p, mu, lambda)
}
\seealso{
\link{wrss}, \link{summation}, \link{constraintL2}, \link{constraintExp2}
}

