% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objClass.R
\name{constraintL2}
\alias{constraintL2}
\title{Soft L2 constraint on parameters}
\usage{
constraintL2(mu, sigma = 1, attr.name = "prior", condition = NULL)
}
\arguments{
\item{mu}{Named numeric vector of prior means.}

\item{sigma}{Named numeric or character vector. Character entries indicate
log-scale sigma parameters to be estimated.}

\item{attr.name}{Character. Name of the attribute storing the constraint value.}

\item{condition}{Optional character vector of conditions.}
}
\value{
Object of class \code{objfn}.
}
\description{
Soft L2 constraint on parameters
}
\details{
Computes
\deqn{(p-\mu)^2 / \sigma^2}
or, if sigma is estimated,
\deqn{(p-\mu)^2 / \sigma^2 + 2\log(\sigma)},
with sigma internally transformed via \code{exp()}.
}
