% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataClass.R
\name{covariates}
\alias{covariates}
\alias{covariates.datalist}
\alias{covariates.data.frame}
\title{Access the covariates in the data}
\usage{
covariates(x, ...)

\method{covariates}{datalist}(x, ...)

\method{covariates}{data.frame}(x, ...)
}
\arguments{
\item{x}{Either a \link{datalist} or a \code{data.frame} with mandatory
columns \code{c("name", "time", "value", "sigma", "lloq")}.}

\item{...}{Additional arguments forwarded to methods (e.g. \code{hypothesis}
for the \code{tbl_df} method).}
}
\value{
The \code{condition.grid} of the data
}
\description{
Access the covariates in the data
}
