% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{createExample}
\alias{createExample}
\title{Open a unit test template.}
\usage{
createExample(test, testPath = NULL)
}
\arguments{
\item{test}{Name of the unit test used as file name with "test-" prefixed.}

\item{testPath}{Unit test folder. Defaults to "inst/tests".}
}
\description{
Open a unit test template.
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}

