% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underConstruction.R
\name{data2forc}
\alias{data2forc}
\title{Translate data into time-continuous data representations}
\usage{
data2forc(data, tau = NULL, ngrid = 1000, type = "linear")
}
\arguments{
\item{data}{\code{data.frame} with name (factor), time (numeric), value (numeric) and sigma (numeric)}

\item{tau}{Numeric. The uncertainty of a single data point is smeared by a Gaussian with variance tau².}

\item{ngrid}{The number of time points returned.}

\item{type}{Character, the type of interpolation, i.e. "linear", "spline", "logspline" or "smooth".}
}
\value{
a \code{data.frame} with columns "name", "time" and "value".
}
\description{
Translate data into time-continuous data representations
}

