% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{datapointL2_indiv}
\alias{datapointL2_indiv}
\title{DatapointL2 without access to stored predictions}
\usage{
datapointL2_indiv(
  name,
  time,
  value,
  sigma = 1,
  attr.name = "validation",
  condition,
  prd_indiv
)
}
\arguments{
\item{name}{character, the name of the prediction, e.g. a state name.}

\item{time}{numeric, the time-point associated to the prediction}

\item{value}{character, the name of the parameter which contains the
prediction value.}

\item{sigma}{numeric, the uncertainty of the introduced test data point}

\item{attr.name}{character. The constraint value is additionally returned in an 
attributed with this name}

\item{condition}{character, the condition for which the prediction is made.}

\item{prd_indiv}{a prediction function}
}
\value{

}
\description{
DatapointL2 without access to stored predictions
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
