% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resObjClass.R
\name{datapointResL2}
\alias{datapointResL2}
\title{L2 objective function for validation data point (residual form)}
\usage{
datapointResL2(name, time, value, sigma = 1, condition)
}
\arguments{
\item{name}{Character, prediction name (e.g., state name)}

\item{time}{Numeric, time point}

\item{value}{Character, parameter name containing the validation value}

\item{sigma}{Numeric, uncertainty}

\item{condition}{Character, condition name}
}
\value{
Object of class `resObjfn`
}
\description{
Creates residual for a single validation point.
Equivalent to `datapointL2()` but returns residual instead of objlist.
}
\seealso{
Other dMod interface: 
\code{\link{constraintResL2}()},
\code{\link{res}()},
\code{\link{resCpp}()},
\code{\link{resL2}()}
}
\concept{dMod interface}
