% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRtools.R
\name{define.IQRpars}
\alias{define.IQRpars}
\title{Define parameterization of IQR model}
\usage{
define.IQRpars(guess, estimate = NULL, transform = NULL, iiv = NULL)
}
\arguments{
\item{guess}{initial guesses for parameters, named numeric.}

\item{estimate}{estimate paramter (1, default) or not (0), named numeric/logical.}

\item{transform}{parameter transformation, no ("N"), log ("L", default), or logit ("G").}

\item{iiv}{parameters to be estimated per individuum, subset of the parameter names. By default none.}
}
\value{
List with guess, estimate, transform and iiv, filled up to the correct lengths.
}
\description{
Define parameterization of IQR model
}
