% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRtools.R
\name{define_IQRpars}
\alias{define_IQRpars}
\title{Define parameterization of IQR model}
\usage{
define_IQRpars(guess, estimate = NULL, transform = NULL, iiv_guess = NULL,
  iiv_estimate = NULL)
}
\arguments{
\item{guess}{initial guesses for parameters, named numeric.}

\item{estimate}{estimate paramter (1, default) or not (0), named numeric/logical.}

\item{transform}{parameter transformation, no ("N"), log ("L", default), or logit ("G").}

\item{iiv_guess}{initial guesses for the omegas of the random effects}

\item{iiv_estimate}{0(default): no random effect, 1: random effect is estimated, 2: random effects are not estimated and take on the value provided in iiv_guess}
}
\value{
List with guess, estimate, transform and iiv, filled up to the correct lengths.
}
\description{
Define parameterization of IQR model
}
