% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classes.R
\name{derivatives}
\alias{derivatives}
\alias{derivatives.list}
\alias{derivatives.objlist}
\alias{derivatives.parvec}
\alias{derivatives.prdframe}
\alias{derivatives.prdlist}
\title{Extract the derivatives of an object}
\usage{
derivatives(x, ...)

\method{derivatives}{parvec}(x, ...)

\method{derivatives}{prdframe}(x, ...)

\method{derivatives}{prdlist}(x, ...)

\method{derivatives}{list}(x, ...)

\method{derivatives}{objlist}(x, ...)
}
\arguments{
\item{x}{object from which the derivatives should be extracted}

\item{...}{additional arguments (not used right now)}
}
\value{
The derivatives in a format that depends on the class of \code{x}.
This is
\code{parvec -> matrix},
\code{prdframe -> prdframe},
\code{prdlist -> prdlist},
\code{objlist -> named numeric}.
}
\description{
Extract the derivatives of an object
}

