% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod.R
\docType{package}
\name{dMod}
\alias{dMod-package}
\alias{dMod}
\title{dMod: Dynamic Modeling in R}
\description{
The framework provides functions to generate ODEs of reaction networks, parameter transformations, observation functions, residual functions, etc. The framework follows the paradigm that derivative information should be used for optimization whenever possible. Therefore, all major functions produce and can handle expressions for symbolic derivatives.
}
\keyword{internal}
