% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{.recon_elim}
\alias{.recon_elim}
\title{Reconstruct eliminated species via compiled function}
\usage{
.recon_elim(recon_eval, root, pv, dependent, parms_all, elim_states)
}
\description{
Reconstruct eliminated species via compiled function
}
\keyword{internal}
