% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqnClass.R
\name{dot}
\alias{dot}
\title{Symbolic time derivative of equation vector given an equation list}
\usage{
dot(observable, eqnlist)
}
\arguments{
\item{observable}{named character vector or object of type \link{eqnvec}}

\item{eqnlist}{equation list}
}
\value{
An object of class \link{eqnvec}
}
\description{
The time evolution of the internal states is defined in the equation list.
Time derivatives of observation functions are expressed in terms of the
rates of the internal states.
}
\details{
Observables are translated into an ODE
}
\examples{
## Generate another equation list
eq <- eqnlist()
eq <- addReaction(eq, "A", "pA", "act_A * A * stimulus", "Phosphorylation of A")
eq <- addReaction(eq, "pA", "A", "deact_A * pA", "Deposphorylation of pA")
eq <- addReaction(eq, "B", "pB", "act_B * B * pA", "Phosphorylation of B")
eq <- addReaction(eq, "pB", "B", "deact_B * pB", "Deposphorylation of pB")

## Extract data.frame of reactions
getReactions(eq)

## Get conserved quantities
conservedQuantities(eq$smatrix)

## Get fluxes
getFluxes(eq)

## Subsetting of equation list
subset(eq, "pB" \%in\% Product)
subset(eq, grepl("Phosphorylation", Description))

## Time derivatives of observables
observables <- eqnvec(pA_obs = "s1*pA", tA_obs = "s2*(A + pA)")
dobs <- dot(observables, eq)

## Combined equation vector for ODE and observables
f <- c(as.eqnvec(eq), dobs)
print(f)
}

