% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/eqnClass.R
\name{eqnvec}
\alias{eqnvec}
\alias{is.eqnvec}
\title{Generate equation vector object}
\usage{
eqnvec(...)

is.eqnvec(x)
}
\arguments{
\item{...}{mathematical expressions as characters to be coerced,
the right-hand sides of the equations}

\item{x}{obect of any class}
}
\value{
object of class \code{eqnvec}, basically a named character.
}
\description{
The eqnvec object stores explicit algebraic equations, like the
right-hand sides of an ODE, observation functions or parameter transformations
as named character vectors.
}
\examples{
v <- eqnvec(y = "2*x + offset")
print(v)
is.eqnvec(v)

}
\seealso{
\link{eqnlist}
}
