% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{exmpextr}
\alias{exmpextr}
\title{Extract an example from a unit test file.}
\usage{
exmpextr(test, testPath = NULL, examplePath = NULL)
}
\arguments{
\item{test}{File name of the test file, source.}

\item{testPath}{Path to test files. Defaults to "inst/tests".}

\item{examplePath}{Path under which the example is stored. Defaults to 
"inst/examples".}
}
\description{
Extract an example from a unit test file.
}
\details{
From \option{test}, an example is extracted and written to a file of
  the same name, but with removed "test-". The file is saved under 
  \option{examplePath}. The special tag #-! can be used to hide code in the 
  test which is enabled again in the example.
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}

