% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{expand.grid.alt}
\alias{expand.grid.alt}
\title{Faster version of expand.grid}
\usage{
expand.grid.alt(seq1, seq2)
}
\arguments{
\item{seq1}{Vector}

\item{seq1}{Vector}
}
\value{
data.frame with the combinations of \code{seq1} and \code{seq2}.
}
\description{
Faster version of expand.grid
}
\details{
See \link{expand.grid} for a description of the functionality.
}

