% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{extractExamples}
\alias{extractExamples}
\title{Extract example from unit tests.}
\usage{
extractExamples(testPath = NULL, examplePath = NULL)
}
\arguments{
\item{testPath}{character, path to test folder}

\item{examplePath}{character path to example folder}
}
\description{
Extract example from unit tests.
}
\details{
Calls exmpextr for all unit test files.
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}

