% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqnClass.R
\name{format.eqnvec}
\alias{format.eqnvec}
\title{Encode equation vector in format with sufficient spaces}
\usage{
\method{format}{eqnvec}(x, ...)
}
\arguments{
\item{x}{object of class \link{eqnvec}. Alternatively, a named parsable character vector.}

\item{...}{additional arguments}
}
\value{
named character
}
\description{
Encode equation vector in format with sufficient spaces
}
