% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R2CdeSolve.R
\name{funC.einspline}
\alias{funC.einspline}
\title{Generate C code for a function and compile it}
\usage{
\method{funC}{einspline}(f, forcings = NULL, jacobian = FALSE,
  boundary = NULL, compile = TRUE, nGridpoints = 500, precision = 1e-05,
  modelname = NULL)
}
\arguments{
\item{f}{named character vector containing the right-hand sides of the ODE}

\item{forcings}{character vector with the names of the forcings}

\item{jacobian}{logical indicating whether the jacobian is computed and written into the C file}

\item{boundary}{data.frame with columns name, yini, yend specifying the boundary condition set-up. NULL if not a boundary value problem}

\item{compile}{logical. If FALSE, only the C file is written}

\item{nGridpoints}{numeric. The number of spline interpolation points}

\item{precision}{numeric. Defines the inner of the time interval, \code{(tmin, tmax) = [tmin+precision, tmax-precision]}}
}
\value{
the name of the generated shared object file together with a number of attributes
}
\description{
Generate C code for a function and compile it
}
\details{
The function replaces variables by arrays \code{y[i]}, etc. and replaces "^" by pow()
in order to have the correct C syntax. The file name of the C-File is derived from \code{f}.
I.e. \code{funC(abc, ...} will generate a file abc.c in the current directory.
Currently, only explicit ODE specification is supported, i.e. you need to have the right-hand sides of the ODE.
In case you have explicitly time-dependent ODEs, please introduce time t as explicit variable, e.g. \code{f <- c(t = "1", x = "a*t - x", ...)}.

The compilation is done via \code{system(R CMD <modelname>.so -leinspline)}. Therefore, make sure einspline is installed
system-wide.
}

