% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateEquations.R
\name{generateEquations}
\alias{generateEquations}
\title{Generate equations from data.frame(s)}
\usage{
generateEquations(...)
}
\arguments{
\item{...}{one or more \code{data.frame}s with columns "Description" (character), "Rate" (character),
and one column per ODE state with the state names. The state columns correspond to the stoichiometric
matrix.}
}
\value{
An object of class \code{eqnList}, a named vector with the equations. Contains attributes "SMatrix"
(the stoichiometric matrix), "species" (the state names), "rates" (the rate expressions) and "description".
}
\description{
Generate equations from data.frame(s)
}
\details{
This function is supposed to translate a reaction network as being defined in a csv file
into the raw equations.
}
\examples{
reactions <- data.frame(Description = c("Activation", "Deactivation"), Rate = c("act*A", "deact*pA"), A=c(-1,1), pA=c(1, -1) )
f <- generateEquations(reactions)
}

