% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toolsDaniel.R
\name{generateModel}
\alias{generateModel}
\title{Generate the model objects for use in Xs (models with sensitivities)}
\usage{
generateModel(f, einspline = FALSE, modelname = "f", ...)
}
\arguments{
\item{f}{Named character vector with the ODE}

\item{einspline}{Logical. Use einspline library if set to \code{TRUE}. Defaults to \code{FALSE}.}

\item{...}{Further arguments being passed to funC.}
}
\value{
list with \code{func} (ODE object) and \code{extended} (ODE+Sensitivities object)
}
\description{
Generate the model objects for use in Xs (models with sensitivities)
}

