% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underConstruction.R
\name{generateModelIE}
\alias{generateModelIE}
\title{Generate the model objects for use in Xv (models with input estimation)}
\usage{
generateModelIE(f, observed, inputs, forcings, scale = 1, modelname = "f",
  ...)
}
\arguments{
\item{f}{Named character vector with the ODE}

\item{observed}{Character vector of the observed states (subset of \code{names(f)})}

\item{inputs}{Character vector of the input function to be estimated as part of the algorithm}

\item{forcings}{Character vector of forcings (the input functions that are NOT estimated
as part of the algorithm but still occur as driving force of the ODE)}

\item{...}{Further arguments being passed to funC.}
}
\value{
list with \code{func_au} (Combined ODE for states and adjoint sensitivities) and 
\code{func_l} (ODE of the log-likelihood and its gradient)
}
\description{
Generate the model objects for use in Xv (models with input estimation)
}

