% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/symbolics.R
\name{getCoefficients}
\alias{getCoefficients}
\title{Get coefficients from a character}
\usage{
getCoefficients(char, symbol)
}
\arguments{
\item{char}{character, e.g. "2*x + y"}

\item{symbol}{single character, e.g. "x" or "y"}
}
\value{
numeric vector with the coefficients
}
\description{
Get coefficients from a character
}
\examples{
getCoefficients("2*x + x + y", "x")
}

