% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{getConditionsSBML}
\alias{getConditionsSBML}
\title{Import condition.grid from PEtab}
\usage{
getConditionsSBML(conditions, data)
}
\arguments{
\item{conditions}{PEtab condition file as .tsv}
}
\value{
condition.grid as data frame.
}
\description{
This function imports the experimental conditions from the PEtab condition file as a gondition.grid.
}
\author{
Marcus Rosenblatt and Svenja Kemmer
}
