% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{getDataPEtabSBML}
\alias{getDataPEtabSBML}
\title{Import Data from PEtab}
\usage{
getDataPEtabSBML(data, observables)

getDataPEtabSBML(data, observables)
}
\arguments{
\item{data}{PEtab data file as .tsv}

\item{observables}{observables as eqnvec}
}
\value{
data as data list and errors (if required) as eqnvec.

data as data list and errors (if required) as eqnvec.
}
\description{
This function imports data from the PEtab data file as a data list and defines errors if an error model is required.

This function imports data from the PEtab data file as a data list and defines errors if an error model is required.
}
\author{
Marcus Rosenblatt and Svenja Kemmer

Marcus Rosenblatt and Svenja Kemmer
}
