% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{getDerivs2}
\alias{getDerivs2}
\alias{getDerivs2.parvec}
\alias{getDerivs2.prdframe}
\alias{getDerivs2.prdlist}
\alias{getDerivs2.list}
\alias{getDerivs2.objlist}
\title{Extract second-order derivatives from an object}
\usage{
getDerivs2(x, ...)

\method{getDerivs2}{parvec}(x, ...)

\method{getDerivs2}{prdframe}(x, ...)

\method{getDerivs2}{prdlist}(x, ...)

\method{getDerivs2}{list}(x, ...)

\method{getDerivs2}{objlist}(x, ...)
}
\arguments{
\item{x}{Object from which the second derivatives should be extracted.
Supported classes are \code{parvec}, \code{prdframe}, \code{prdlist}, \code{list}, and
\code{objlist}.}

\item{...}{Additional arguments passed to specific methods (currently unused).}
}
\value{
The structure of the returned object depends on the class of \code{x}:
\itemize{
\item \code{parvec} – a 3D array \verb{[p, theta, theta]} of second derivatives.
\item \code{prdframe} – a 4D array \verb{[time, variable, theta, theta]}.
\item \code{prdlist} – a list of \code{prdframe} second-derivative arrays.
\item \code{objlist} – the stored \code{hessian} matrix.
\item \code{list} – a list of derivative objects, depending on the elements.
}
}
\description{
Generic accessor for the \code{deriv2} attribute (or \code{hessian} field, in the
case of \code{objlist}) attached to dMod objects. Mirrors \link{getDerivs} for
first-order derivatives.
}
\examples{
\dontrun{
d2 <- getDerivs2(myprdframe)
getDerivs2(myparvec)
}

}
