% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{getEstGridParameterMapping}
\alias{getEstGridParameterMapping}
\title{Get Parameter mappings outer.estgrid - inner.estgrid}
\usage{
getEstGridParameterMapping(x)
}
\arguments{
\item{x}{est.grid}
}
\value{
named character
}
\description{
Get Parameter mappings outer.estgrid - inner.estgrid
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
