% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{getInitialsSBML}
\alias{getInitialsSBML}
\title{Import initials from SBML.}
\usage{
getInitialsSBML(model, conditions)
}
\arguments{
\item{model}{SBML file as .xml}
}
\value{
Named vector of initials.
}
\description{
This function imports initial values or equations describing the same from SBML and writes them in a named vector.
}
\author{
Marcus Rosenblatt, Svenja Kemmer and Frank Bergmann
}
