% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqnClass.R
\name{getLinVars}
\alias{getLinVars}
\title{Identify linear variables in an equation vector using sympy}
\usage{
getLinVars(eqnvec)
}
\arguments{
\item{eqnvec}{An object of class \code{eqnvec}, representing a set of equations.}
}
\value{
A character vector of variables that occur linearly in all equations.
}
\description{
Identify linear variables in an equation vector using sympy
}
\details{
This function calls Python's `sympy` library via `reticulate` to symbolically analyze equations and determine if variables appear linearly in all equations.
}
\examples{
eqnvec <- as.eqnvec(
  c("-k1*A", "k1*A - k2*B", "-k3*B*C/(Km+C) + k4*pC", "k3*B*C/(Km+C) - k4*pC"),
  names = c("A", "B", "C", "pC")
)
getLinVars(eqnvec)

}
