% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getLocalDLLs}
\alias{getLocalDLLs}
\title{Determine loaded DLLs available in working directory}
\usage{
getLocalDLLs()
}
\value{
Character vector with the names of the loaded DLLs available in the working directory
}
\description{
Determine loaded DLLs available in working directory
}
