% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{getObservables}
\alias{getObservables}
\alias{getObservables.tbl_df}
\title{Extract the observables of an object}
\usage{
getObservables(.x, ...)

\method{getObservables}{tbl_df}(dMod.frame, hypothesis = 1)
}
\arguments{
\item{hypothesis}{The hypothesis in the dMod.frame}

\item{x}{object from which the equations should be extracted}
}
\value{
The equations as a character.
}
\description{
Extract the observables of an object
}
