% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{getObservablesSBML}
\alias{getObservablesSBML}
\title{Import observables from PEtab.}
\usage{
getObservablesSBML(observables)
}
\arguments{
\item{observables}{PEtab observable file as .tsv}
}
\value{
Eqnvec of observables.
}
\description{
This function imports observables from the PEtab observable file.
}
\author{
Marcus Rosenblatt and Svenja Kemmer
}
