% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{getParameters.data.table}
\alias{getParameters.data.table}
\title{Title}
\usage{
\method{getParameters}{data.table}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{

}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
