% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{getParameters2.data.table}
\alias{getParameters2.data.table}
\title{Get Parameter mappings outer.estgrid - inner.estgrid}
\usage{
getParameters2.data.table(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{

}
\description{
Get Parameter mappings outer.estgrid - inner.estgrid
}
