% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{getParameters_est.grid}
\alias{getParameters_est.grid}
\title{Extract parameter names from est.grid}
\usage{
getParameters_est.grid(est.grid)
}
\arguments{
\item{est.grid}{data.frame(parname, partask, ids...)}
}
\value{
character of outer parameter names
}
\description{
Extract parameter names from est.grid
}
