% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{getParametersSBML}
\alias{getParametersSBML}
\title{Import Parameters from PEtab}
\usage{
getParametersSBML(parameters, model)
}
\arguments{
\item{parameters}{PEtab parameter file as .tsv}
}
\value{
constraints and pouter as list of named vectros.
}
\description{
This function imports fixed and fitted parameters from the PEtab parameter file as named vectors.
}
\author{
Marcus Rosenblatt and Svenja Kemmer
}
