% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{getParametersToEstimate}
\alias{getParametersToEstimate}
\title{Determine which parameters need sensitivities}
\usage{
getParametersToEstimate(est.grid, trafo, reactions)
}
\arguments{
\item{est.grid}{est.grid = data.table(condition, par1,...,parN)}

\item{trafo}{symbolic base trafo}

\item{reactions}{Object of class \code{\link[=eqnlist]{eqnlist()}}}
}
\value{
character, to go into the estimate-argument of odemodel
}
\description{
Determine which parameters need sensitivities
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
