% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod2PEtab_Svenja.R
\name{getParGrids}
\alias{getParGrids}
\title{Get est.grid and fixed.grid}
\usage{
getParGrids(
  mytrafo = trafo,
  mytrafoL = trafoL,
  mycondition.grid = condition.grid,
  SS_pars = names(mySS_eqns_est)
)
}
\arguments{
\item{jobname}{job name on cluster}

\item{recover}{}

\item{nodetype}{}
}
\value{

}
\description{
Get est.grid and fixed.grid
}
