% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{getReactionsSBML}
\alias{getReactionsSBML}
\title{Import reactions from SBML.}
\usage{
getReactionsSBML(model, conditions)
}
\arguments{
\item{model}{SBML file as .xml}
}
\value{
Eqnlist of reactions.
}
\description{
This function imports reactions from SBML. Reactions are written to an eqnlist object. 
Assignment rules for input functions are substituted in the rate column. Time is introduced as an additionel state t.
}
\author{
Marcus Rosenblatt and Svenja Kemmer
}
