% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{getTrafoType}
\alias{getTrafoType}
\title{Determine the type of trafo for each element of a vector}
\usage{
getTrafoType(trafo_string)
}
\arguments{
\item{trafo_string}{named character}
}
\value{
character(lenth(trafo_string)):
TRAFO = functional trafo
SYMBOL = Parameter name mapping
NUMBER = fixed value
}
\description{
Determine the type of trafo for each element of a vector
}
\examples{
trafo_string <- c(STAT5A = "207.6 * ratio", STAT5B = "207.6 - 207.6 * ratio", 
                  pApB = "alpha", pApA = "beta", pBpB = "pBpB", nucpApA = "0.1", nucpApB = "0", 
                  nucpBpB = "0")
getTrafoType(trafo_string)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
