% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toolsMirjam.R
\name{getZeroFluxesInt}
\alias{getZeroFluxesInt}
\title{Find fluxes with low contributions}
\usage{
getZeroFluxesInt(out, rtol = 0.05, atol = 0)
}
\arguments{
\item{out}{list of integrated fluxes, named "state.fluxEquation". States are not allowed to contain a "."-symbol.}
}
\description{
Find fluxes with low contributions
}
\details{
This can be used if the fluxes were integrated (by using \link{Xf} to integrate the fluxes),
otherwise see \link{getZeroFluxes}.
}

