% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDaniel.R
\name{ggdata}
\alias{ggdata}
\title{Produce data frames for plotting}
\usage{
ggdata(prediction = NULL, data = NULL, ..., condition.grid = attr(data,
  "condition.grid"), substitute = "_")
}
\arguments{
\item{prediction}{object of class prediction list}

\item{data}{object of class data list}

\item{...}{arguments going to subsetting of prediction and data}

\item{condition.grid}{data.frame with row.names according to the condition names.}

\item{substitute}{Substitute "_" by character \code{substitute}.}
}
\value{
list with prediction (data.frame) and data (data.frame)
}
\description{
Produce data frames for plotting
}
