% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDaniel.R
\name{ggdata_fn}
\alias{ggdata_fn}
\title{Produce data frames for plotting}
\usage{
ggdata_fn(prdfn = NULL, errfn = NULL, data = NULL, times, pars, ...,
  condition.grid = attr(data, "condition.grid"), substitute = "_")
}
\arguments{
\item{data}{object of class data list}

\item{...}{arguments going to subsetting of prediction and data}

\item{condition.grid}{data.frame with row.names according to the condition names.}

\item{substitute}{Substitute "_" by character \code{substitute}.}

\item{prediction}{object of class prediction list}
}
\value{
list with prediction (data.frame) and data (data.frame)
}
\description{
Produce data frames for plotting
}
