% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{ginv}
\alias{ginv}
\title{Generalized Inverse of a Matrix}
\usage{
ginv(X, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{Matrix for which the Moore-Penrose inverse is required.}

\item{tol}{A relative tolerance to detect zero singular values.}
}
\value{
A MP generalized inverse matrix for X.
}
\description{
Calculates the Moore-Penrose generalized inverse of a matrix X.
}
\concept{Auxiliary}
