% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModFrameClass.R
\name{git_add_dMod.frame}
\alias{git_add_dMod.frame}
\title{Stage a dMod.frame and all of its DLLs}
\usage{
git_add_dMod.frame(dMod.frame)
}
\arguments{
\item{dMod.frame}{the dMod.frame or a character vector specifying a RDS-file}
}
\value{
This function is called for its side-effects.
}
\description{
Stage a dMod.frame and all of its DLLs
}
