% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{\%.*\%.resObjlist}
\alias{\%.*\%.resObjlist}
\title{Multiplication of residual objective functions with scalars}
\usage{
`\%.*\%.resObjlist`(x1, x2)
}
\arguments{
\item{x1}{Numeric scalar}

\item{x2}{Object of class resObjfn or resObjlist}
}
\value{
A residual objective function resObjfn or resObjlist object
}
\description{
The \code{\%.*\%} operator allows to multiply objects of class 
resObj or resObjfn with a scalar.
}
