% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{\%.*\%}
\alias{\%.*\%}
\title{Multiplication of objective functions with scalars}
\usage{
x1 \%.*\% x2
}
\arguments{
\item{x1}{object of class objfn or objlist.}

\item{x2}{numeric of length one.}
}
\value{
An objective function or objlist object.
}
\description{
The \code{\%.*\%} operator allows to multiply objects of class objlist or objfn with
a scalar.
}
