% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{gridlist}
\alias{gridlist}
\title{Collect grids in list}
\usage{
gridlist(est.grid, fix.grid)
}
\arguments{
\item{est.grid}{}

\item{fix.grid}{}
}
\value{

}
\description{
Ensure all tables are data.tables
}
