% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{Id}
\alias{Id}
\title{An identity function which vanishes upon concatenation of fns}
\usage{
Id()
}
\value{
fn of class idfn
}
\description{
An identity function which vanishes upon concatenation of fns
}
\examples{
x <- Xt()
id <- Id()

(id*x)(1:10, pars = c(a = 1))
(x*id)(1:10, pars = c(a = 1))
str(id*x)
str(x*id)
}
