% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{importPEtabSBML}
\alias{importPEtabSBML}
\title{Import an SBML model and corresponding PEtab objects}
\usage{
importPEtabSBML(
  modelname = "Boehm_JProteomeRes2014",
  path2model = "BenchmarkModels/",
  testCases = FALSE,
  path2TestCases = "PEtabTests/",
  compile = TRUE,
  SBML_file = NULL,
  observable_file = NULL,
  condition_file = NULL,
  data_file = NULL,
  parameter_file = NULL
)
}
\arguments{
\item{modelname}{name of folder containing all PEtab files of the model to be imported. NULL if file paths are defined separately (see below).}

\item{path2model}{path to model folder}

\item{path2TestCases}{path to feature test case folder}

\item{compile}{if FALSE, g, ODEmodel and err are loaded from .RData (if present) and compilation time is saved}

\item{SBML_file}{SBML model as .xml}

\item{observable_file}{PEtab observable file as .tsv}

\item{condition_file}{PEtab condition file as .tsv}

\item{data_file}{PEtab data file as .tsv}

\item{parameter_file}{PEtab parameter file as .tsv}

\item{TestCases}{TRUE to load feature test cases}
}
\value{
name of imported model
}
\description{
This function imports an SBML model and corresponding PEtab files, e.g. from the Benchmark collection.
}
\details{
Objects such as model equations, parameters or data are automatically assigned to the following standard variables and written to your current working directory (via <<-):
reactions, observables, errors, g, x, p0, err, obj, mydata, ODEmodel, condition.grid, trafoL, pouter, times.
Compiled objects (g, ODEmodel and err) are saved in .RData.
}
\author{
Marcus Rosenblatt and Svenja Kemmer
}
