% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{importPEtabSBML_indiv}
\alias{importPEtabSBML_indiv}
\title{Import an SBML model and corresponding PEtab objects}
\usage{
importPEtabSBML_indiv(
  filename = "enzymeKinetics/enzymeKinetics.petab",
  testCases = FALSE,
  path2TestCases = "PEtabTests/",
  NFLAGcompile = c(Recompile = 0, RebuildGrids = 1, LoadPrevious = 2)[3],
  SFLAGbrowser = c(NA, "Beginning", "BuildGrids", "Compilation", "CollectList")[1]
)
}
\arguments{
\item{filename}{"path/to/modelname.petab". Will look for files like
"path/to/modelname/model_modelname.xml"}

\item{path2TestCases}{path to feature test case folder}

\item{TestCases}{TRUE to load feature test cases}

\item{compile}{if FALSE, g, ODEmodel and err are loaded from .RData (if present) and compilation time is saved}
}
\value{
list of dMod model objects with entries
\itemize{
\item symbolicEquations
\item odemodel
\item data
\item gridlist
\item e
\item fns
\item prd
\item obj_data
\item pars
}
}
\description{
This function imports an SBML model and corresponding PEtab files, e.g. from the Benchmark collection.
}
\details{
Objects such as model equations, parameters or data are automatically assigned to the following standard variables and written to your current working directory (via <<-):
reactions, observables, errors, g, x, p0, err, obj, mydata, ODEmodel, condition.grid, trafoL, pouter, times.
Compiled objects (g, ODEmodel and err) are saved in .RData.
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de) building on the original function of Marcus and Svenja
}
