% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{indiv2Classic}
\alias{indiv2Classic}
\title{Title}
\usage{
indiv2Classic(
  pd,
  .compiledFolder = file.path("CompiledObjects"),
  Nobjtimes = 50
)
}
\arguments{
\item{pd}{}

\item{.compiledFolder}{}
}
\value{
a `pd` object but with classic fns instead of indiv fns
}
\description{
Title
}
