% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{indiv2classic_getFasterVersion}
\alias{indiv2classic_getFasterVersion}
\title{Get the faster version of classic and indiv}
\usage{
indiv2classic_getFasterVersion(pd, NFLAGcompile = 1)
}
\arguments{
\item{pd}{pd_indiv}

\item{NFLAGcompile}{recompile the parameter trafo of pdc?}
}
\value{

}
\description{
Get the faster version of classic and indiv
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
