% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{indiv2Classic_merge_grids}
\alias{indiv2Classic_merge_grids}
\title{Title}
\usage{
indiv2Classic_merge_grids(eg, fg)
}
\arguments{
\item{eg}{est.grid}

\item{fg}{fix.grid}
}
\value{
condition.grid(condition, parsMixedCols...)
}
\description{
Title
}
