% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{indiv_addLocalParsToGridList}
\alias{indiv_addLocalParsToGridList}
\title{Add individualized parameters to grids}
\usage{
indiv_addLocalParsToGridList(pars, gridlist, FLAGoverwrite = FALSE)
}
\arguments{
\item{pars}{data.table(ID, condition, pars...) Only one of ID, condition must be present}

\item{gridlist}{\code{\link[=gridlist]{gridlist()}}}
}
\value{
modified gridlist
}
\description{
Can only add parameters, cannot update existing parameters
}
\examples{
pars <- data.table(ID = 1:2, newFix = c(1,2), newEst = c("a", "b"), newMix = c("a", 1))
est.grid <- data.table(ID = 1:2,
                       condition = c("A", "B"),
                       k1 = c("k1_A", "k1_B"),
                       k2 = c("k2_A", "k2_B"),
                       k3 = c("k3", NA),
                       k4 = c("k4", "k4"),
                       stringsAsFactors = FALSE)
fix.grid <- data.table(ID = 1:2,
                       condition = c("A", "B"),
                       k3 = c(NA, 3.5),
                       k5 = c(5.1,5.2),
                       k6 = c(6,6),
                       stringsAsFactors = FALSE)
gl <- gridlist(est.grid = est.grid, fix.grid = fix.grid)
indiv_addLocalParsToGridList(pars, gl)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
