% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{init_empty_objlist}
\alias{init_empty_objlist}
\title{Create an objlist with zeros as entries}
\usage{
init_empty_objlist(pars, deriv = TRUE)
}
\arguments{
\item{pars}{named vector. Only names and length are used}

\item{deriv}{TRUE or FALSE}
}
\description{
Create an objlist with zeros as entries
}
\examples{
init_empty_objlist(setNames(rnorm(5), letters[1:5]))
}
