% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDenial.R
\name{inputFileChanged}
\alias{inputFileChanged}
\title{Title}
\usage{
inputFileChanged(filename_input, filename_output)
}
\arguments{
\item{filename_input}{}

\item{filename_output}{}
}
\value{
TRUE if inputfile is newer than outputfile or outputfile does not exist
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
