% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toolsMirjam.R
\name{intersectList}
\alias{intersectList}
\title{Find common elements in lists}
\usage{
intersectList(list, byNames = TRUE)
}
\arguments{
\item{list}{contains a set of vectors.}

\item{byNames}{if set TRUE common names are checked instead of common values (default: TRUE)}
}
\description{
generalisation of \link{intersect} to a list of vectors.
}
\examples{
testList <-list(c(a=1,b=5,c=3,e=5), c(d=3,b=1,q=1,c=5,i=2))
intersectList(testList,FALSE)
intersectList(testList,TRUE)
}

