% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRtools.R
\name{IQRdMod.frame}
\alias{IQRdMod.frame}
\title{Create dMod.frame from IQRtools imports}
\usage{
IQRdMod.frame(hypothesis = date(), model, data, parameters)
}
\arguments{
\item{hypothesis}{hypothesis name}

\item{model}{result from \link{read.IQRmodel()}}

\item{data}{result from \link{read.IQRdata()}}

\item{parameters}{result from \link{define.IQRpars()}}
}
\value{
A \link{dMod.frame} object
}
\description{
Create dMod.frame from IQRtools imports
}
