% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRtools.R
\name{IQRsysProject}
\alias{IQRsysProject}
\title{Create a dMod.frame from IQRtools imports}
\usage{
IQRsysProject(model, data, parameters, projectPath = NULL)
}
\arguments{
\item{model}{result from \link{read.IQRmodel()}}

\item{data}{result from \link{read.IQRdata()}}

\item{parameters}{result from \link{define.IQRpars()}}

\item{projectPath}{the path where to generate the project and store the results}
}
\value{
A \link{dMod.frame} object
}
\description{
Create a dMod.frame from IQRtools imports
}
