% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{is.objframe}
\alias{is.objframe}
\title{Check if object is an objframe}
\usage{
is.objframe(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Test whether an object is of class objframe
}
