% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objClass.R
\name{is.objlist}
\alias{is.objlist}
\title{Tests for objlist}
\usage{
is.objlist(x)
}
\arguments{
\item{x}{objlist}
}
\value{
T or F
}
\description{
Tests if value is numeric, gradient is named numeric, hessian is numeric matrix with dimnames. 
NULL is allowed for each of the three
}
\examples{
is.objlist(objlist(1,1,2))
}
