% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsSeverin.R
\name{julia_setup}
\alias{julia_setup}
\title{Install the julia setup}
\usage{
julia_setup(installJulia = FALSE, installJuliaPackages = TRUE)
}
\arguments{
\item{installJulia}{boolean, default \code{false}. If set to true, juliaup and via this then Julia is installed.}

\item{installJuliaPackages}{boolean, default \code{true}. If set to true, the necessary packages are installed.}
}
\value{
nothing
}
\description{
Installs Julia and the necessary julia packages
}
