% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trustL1.R
\name{l1norm}
\alias{l1norm}
\title{Soft L1 prior on parameters}
\usage{
l1norm(nout)
}
\arguments{
\item{nout}{data.frame (result of \link{res})}

\item{p}{Namec numeric, the parameter value}

\item{mu}{Named numeric, the prior values}

\item{lambda}{Named numeric of length of mu or numeric of length one.}

\item{fixed}{Named numeric with fixed parameter values (contribute to the prior value
but not to gradient and Hessian)}
}
\value{
List of class \code{obj}, i.e. objective value, gradient and Hessian as list.

list with entries value (numeric, the weighted residual sum of squares), 
gradient (numeric, gradient) and 
hessian (matrix of type numeric).
}
\description{
Soft L1 prior on parameters
}
\details{
Computes the constraint value 
\deqn{\lambda\|p-\mu\|}{lambda*abs(p-mu)}
and its derivatives with respect to p.
}
\examples{
p <- c(A = 1, B = 2, C = 3)
mu <- c(A = 0, B = 0)
lambda <- c(A = 0.1, B = 1)
constraintL1(p, mu, lambda)
Compute the L1 norm of residuals

}
\seealso{
\link{wrss}, \link{summation}, \link{constraintL2}, \link{constraintExp2}
}
