% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{lbind}
\alias{lbind}
\title{Bind named list of data.frames into one data.frame}
\usage{
lbind(mylist)
}
\arguments{
\item{mylist}{A named list of data.frame. The data.frames are expected to have the same structure.}
}
\value{
data.frame with the originial columns augmented by a "condition" column.
}
\description{
Bind named list of data.frames into one data.frame
}
\details{
Each data.frame ist augented by a "condition" column containing the name attributed of
the list entry. Subsequently, the augmented data.frames are bound together by \code{rbind}.
}
