% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R2CdeSolve.R
\name{loadDLL}
\alias{loadDLL}
\title{Dynamically load DLL with automatic unloading of conflicting DLLs}
\usage{
loadDLL(func, cfunction)
}
\arguments{
\item{func}{result from funC(), contains the information about the DLL name to be loaded}

\item{cfunction}{character, denoting the C function name.}
}
\description{
Dynamically load DLL with automatic unloading of conflicting DLLs
}
\details{
If the C function name is already part of another loaded DLL, the corresponding DLL is
unloaded and the desired func DLL is loaded instead.
}

