% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{loaddMod}
\alias{loaddMod}
\title{Load dMod snapshot}
\usage{
loaddMod(filename = stop("'filename' must be specified"), flags = NULL)
}
\arguments{
\item{filename}{filename, expectes a zip file produced by 
\link{savedMod}}

\item{flags}{additional flags going to \code{R CMD SHLIB}, e.g.
\code{"-leinspline"}.}
}
\value{
Saved objects are loaded into the global environment. 
C files contained in the snapshot file are compiled with additional
flags if necessary. An objective function is created from dMod
objects.
}
\description{
Load dMod snapshot
}
\seealso{
\link{savedMod}
}

