% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{loadTemplate}
\alias{loadTemplate}
\title{Load a template file in the editor}
\usage{
loadTemplate(i = 1)
}
\arguments{
\item{i}{Integer, choose a template to be loaded}
}
\description{
Load a template file in the editor
}
\details{
Possible templates are:
i = 1: Do parameter estimation in a dynamic model with fixed forcings
}

