% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{long2wide}
\alias{long2wide}
\title{Translate long to wide format (inverse of wide2long.matrix)}
\usage{
long2wide(out)
}
\arguments{
\item{out}{data.frame in long format}
}
\value{
data.frame in wide format
}
\description{
Translate long to wide format (inverse of wide2long.matrix)
}

