% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toolsMirjam.R
\name{lsdMod}
\alias{lsdMod}
\title{Print list of dMod objects in .GlobalEnv}
\usage{
lsdMod(classlist = c("objfn", "prdfn"), envir = .GlobalEnv)
}
\arguments{
\item{classlist}{List of object classes to print.}

\item{envir}{Alternative environment to search for objects.}
}
\description{
Lists the objects for a set of classes.
}
\examples{
\dontrun{
lsdMod()
lsdMod(classlist = "prdfn", envir = environment(obj))
}
}

