% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{make_pars}
\alias{make_pars}
\title{Extract individualized parameters for one ID}
\usage{
make_pars(parsouter, fixedouter, condition, est.grid, fixed.grid)
}
\arguments{
\item{est.grid}{}
}
\value{
c(par0 = value)
}
\description{
Catch the following two cases:
* Entry in est.grid is NA                   => Parameter value present in fixed.grid
* Parameter name not present in parsouter => Parameter value present in fixed_outer
}
