% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{match.fnargs}
\alias{match.fnargs}
\title{dMod match function arguments}
\usage{
match.fnargs(arglist, choices)
}
\arguments{
\item{arglist}{list}

\item{choices}{character}
}
\description{
The function is exported for dependency reasons
}
