% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mname}
\alias{mname}
\alias{mname.NULL}
\alias{mname.character}
\alias{mname.objfn}
\alias{mname.fn}
\title{Get modelname from single object (used internally)}
\usage{
mname(x, conditions = NULL)

\method{mname}{NULL}(x, conditions = NULL)

\method{mname}{character}(x, conditions = NULL)

\method{mname}{objfn}(x, conditions = NULL)

\method{mname}{fn}(x, conditions = NULL)
}
\arguments{
\item{x}{dMod object}

\item{conditions}{character vector of conditions}
}
\description{
Get modelname from single object (used internally)
}
