% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{modelname}
\alias{modelname}
\alias{modelname<-}
\alias{modelname<-.fn}
\alias{modelname<-.objfn}
\title{Get and set modelname}
\usage{
modelname(..., conditions = NULL)

modelname(x, ...) <- value

\method{modelname}{fn}(x, conditions = NULL, ...) <- value

\method{modelname}{objfn}(x, conditions = NULL, ...) <- value
}
\arguments{
\item{...}{objects of type \code{prdfn}, \code{parfn}, \code{objfn}}

\item{conditions}{character vector of conditions}

\item{x}{dMod object for which the model name should be set}

\item{value}{character, the new modelname (does not change the C file)}
}
\value{
character vector of model names, corresponding to C files
in the local directory.
}
\description{
The modelname attribute refers to the name of a C file associated with
a dMod function object like prediction-, parameter transformation- or
objective functions.
}
