% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toolsWolfgang.R
\name{msbest}
\alias{msbest}
\title{Select best fit.}
\usage{
msbest(fitlist)
}
\arguments{
\item{fitlist}{A data frame of fits as returned from \code{\link{mstrust}}.
       The data frame does not need to be ordered and can include unconverged
       fits.}
}
\value{
The converged fit with lowest chisquare as a named numeric vector.
}
\description{
Select the fit with lowest chi^2 form the result of \code{\link{mstrust}}.
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}

