% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsWolfgang.R
\name{msnarrow}
\alias{msnarrow}
\title{Discover promissing regions in parameter space}
\usage{
msnarrow(center, spread, fits = 100, safety = "", ...)
}
\arguments{
\item{center}{Parameter values for the first ms fit.}

\item{spread}{Vector giving the spread of initial parameters for consecutive
ms fits. The length of <spread> defines how many consecutive ms fits
are run.}

\item{fits}{Number of fits carried out per ms fit. If <fits> is a integer the
same number of fits are used for all ms fits. If a different number of
fits are desired for each ms fit, <fits> must be a vector of the same
length as <spread>.}

\item{safety}{If a non-empty string is given, the fitlist of each mstrust is
written to a file <safety> with the number of the current run
appended.}

\item{...}{Parameters handed to the mulit start optimizer. Right now
\code{\link{mstrust}} is the only option available.}
}
\value{
fitlist A data frame of the fits of the last multi start fit.
}
\description{
Use consecutive multi start (ms) fits to arrive in a well-behaved region in
parameter space.
}
\details{
It might be desirable to start a multi start (ms) fit in a region in
parameter space yielding reasonable fit results. If such a region is unknown,
it can be hoped, that the best fit of a ms fit with a large spread of inital
values identifies such a region. In order to explore the region in greater
detail, a consecutive ms fit about the best fit can be used. The procedure of
consecutive ms-fits about the best fit of the last can be iterated. Thus, it
is expected to arrive in a region in parameter space which gives reasonable
fits.

In case any of the ms fits return without a single valid result, the
algorithm stops, and the fitlist of the last ms fit is returned. Moreover, a
warning is issued. In case the first fit failes in such a way, NULL is
returned.
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}
