% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDaniel.R
\name{mssample}
\alias{mssample}
\title{Generate sample for multi-start fit}
\usage{
mssample(center, samplefun = "rnorm", fits = 20, ...)
}
\arguments{
\item{center}{named numeric, the center around we sample}

\item{samplefun}{character, indicating the random number generator,
defaults to \code{"rnorm"}.}

\item{fits}{length of the sample}

\item{...}{arguments going to \code{samplefun}}
}
\value{
matrix with the parameter samples
}
\description{
Generate sample for multi-start fit
}

