% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objClass.R
\name{nll}
\alias{nll}
\title{Compute the negative log-likelihood}
\usage{
nll(
  nout,
  pars,
  deriv,
  opt.BLOQ = "M3",
  opt.hessian = c(ALOQ_part1 = TRUE, ALOQ_part2 = TRUE, ALOQ_part3 = TRUE, BLOQ_part1 =
    TRUE, BLOQ_part2 = TRUE, BLOQ_part3 = TRUE, PD = TRUE),
  bessel.correction = 1
)
}
\arguments{
\item{nout}{data.frame (result of \link{res}) or object of class \link{res}.}

\item{pars}{Named vector of ALL outer parameters (union across conditions)}

\item{deriv}{Logical. If TRUE, compute gradient and hessian}

\item{opt.BLOQ}{Character denoting the method to deal with BLOQ data.
One of "M1", "M3", "M4NM", or "M4BEAL".}

\item{opt.hessian}{Named logical vector to include or exclude various
summands of the hessian matrix.}

\item{bessel.correction}{Numeric. Bessel correction factor for variance estimation.}
}
\value{
list with entries value, gradient, and hessian (Gauss-Newton approximation).
}
\description{
Gaussian Log-likelihood. Supports NONMEM-like BLOQ handling methods M1, M3 and M4
and estimation of error models with optional Bessel correction for variance parameter bias.
The Hessian is approximated via the Jacobian of the residuals (Gauss-Newton approximation).
Supports different parameter sets per condition - gradients and Hessians are merged by parameter name.
}
