% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objClass.R
\name{nll_BLOQ}
\alias{nll_BLOQ}
\title{Non-linear log likelihood for the BLOQ part of the data}
\usage{
nll_BLOQ(
  nout.bloq,
  derivs.bloq,
  derivs.err.bloq,
  derivs2 = NULL,
  derivs2.err = NULL,
  par_names,
  opt.BLOQ = c("M3", "M4NM", "M4BEAL", "M1"),
  opt.hessian = c(BLOQ_part1 = TRUE, BLOQ_part2 = TRUE, BLOQ_part3 = TRUE)
)
}
\arguments{
\item{nout.bloq}{The bloq output of \code{\link[=res]{res()}}}

\item{derivs.bloq, derivs.err.bloq}{matrix of first derivatives}

\item{derivs2, derivs2.err}{Optional 3D arrays of second prediction and
error-model derivatives over the BLOQ rows. When non-\code{NULL}, the exact
d^2 pred / d^2 sigma contributions to the Hessian are added on top of
the Gauss-Newton-plus-Parts form.}

\item{par_names}{Character vector of ALL parameter names (full set)}

\item{opt.BLOQ}{Character denoting the method to deal with BLOQ data}

\item{opt.hessian}{Named logical vector for hessian components}
}
\description{
Non-linear log likelihood for the BLOQ part of the data
}
