% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{normIndiv}
\alias{normIndiv}
\title{Title}
\usage{
normIndiv(data, prd0, errmodel = NULL, forcings = NULL, iiv = NULL,
  conditional = NULL, fixed.grid, nauxtimes = 500, cores = 1,
  deriv = TRUE, attr.name = "data")
}
\arguments{
\item{iiv}{c("ETA_EC50", "ETA_EMAX")}

\item{conditional}{data.frame}

\item{attr.name}{}
}
\description{
Title
}
