% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{normL2_indiv}
\alias{normL2_indiv}
\title{Fast normL2}
\usage{
normL2_indiv(
  data,
  prd0,
  errmodel = NULL,
  est.grid,
  fix.grid,
  times = NULL,
  attr.name = "data",
  mycores = 1
)
}
\arguments{
\item{data}{}

\item{prd0}{}

\item{errmodel}{}

\item{est.grid}{}

\item{fix.grid}{}

\item{times}{}

\item{attr.name}{}

\item{mycores}{You can use mycores > 1 for parallel computation of conditions to speed up single fits or profiles. Use carefully when also performing multi-start optimization via mstrust.}
}
\value{
objective function
}
\description{
Fast normL2
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
