% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{objframe}
\alias{objframe}
\title{Objective frame}
\usage{
objframe(mydata, deriv = NULL, deriv.err = NULL)
}
\arguments{
\item{mydata}{data.table produced by \link{res}}

\item{deriv}{numeric matrix of first-order derivatives of residuals (Jacobian)}

\item{deriv.err}{numeric matrix of first-order derivatives of the error model}
}
\value{
An object of class \code{"objframe"} (data.table) with attributes \code{"deriv"} and \code{"deriv.err"}.
These arrays have the same parameter axes as those returned by \link{prdframe} and \link{res}.
}
\description{
An objective frame stores residuals and their derivatives with respect to parameters.
It is typically created by \link{res} and used internally in objective functions.
}
