% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDenial.R
\name{objtimes}
\alias{objtimes}
\title{Get vector of logarithmically spaced time points}
\usage{
objtimes(datatimes, eventtimes = NULL, Nobjtimes = 25)
}
\arguments{
\item{datatimes}{times present in data}

\item{eventtimes}{times present in events (not yet implemented)}
}
\value{
vector of times, including datatimes
}
\description{
to help the objective function integrator
}
\examples{
objtimes(c(0,30,60,90,600))
objtimes(c(30,60,90,600))
objtimes(c(-30,60,90,600))
}
