% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R2CdeSolve.R
\name{odeC.einspline}
\alias{odeC.einspline}
\title{Interface to ode()}
\usage{
\method{odeC}{einspline}(y, times, func, parms, ...)
}
\arguments{
\item{y}{named vector of type numeric. Initial values for the integration}

\item{times}{vector of type numeric. Integration times}

\item{func}{return value from funC()}

\item{parms}{named vector of type numeric.}

\item{...}{further arguments going to \code{ode()}}

\item{method}{character, indicating the integration method}
}
\value{
matrix with times and states
}
\description{
Interface to ode()
}
\details{
See deSolve-package for a full description of possible arguments
}

