% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{optimize}
\alias{optimize}
\title{Optimize an Objective Function using Various Optimization Methods using optimx}
\usage{
optimize(
  objfn,
  parinit,
  method = "L-BFGS-B",
  lower = -Inf,
  upper = Inf,
  control = list(),
  ...
)
}
\arguments{
\item{objfn}{A function that computes and returns a list with components `value`, `gradient`, and `hessian`.
This represents the objective function to be minimized.}

\item{parinit}{A numeric vector of initial parameter values.}

\item{method}{A character string specifying the optimization method. Defaults to "L-BFGS-B".
Available methods include those supported by the `optimx` package, such as "BFGS",
"Nelder-Mead", "L-BFGS-B", etc.}

\item{lower}{A numeric vector of lower bounds for the parameters (used only by methods that support
box constraints, e.g., "L-BFGS-B"). Defaults to `-Inf`.}

\item{upper}{A numeric vector of upper bounds for the parameters (used only by methods that support
box constraints, e.g., "L-BFGS-B"). Defaults to `Inf`.}

\item{control}{A list of control parameters to pass to the optimization algorithm.}

\item{...}{Additional arguments to pass to the objective function.}
}
\value{
A list containing:
        - `value`: The value of the objective function at the optimum.
        - `gradient`: The gradient at the optimum.
        - `hessian`: The Hessian at the optimum.
        - `argument`: The optimized parameters.
        - `converged`: Logical indicating if the optimizer converged.
        - `iterations`: The number of function evaluations.
}
\description{
This function extends the existing optimization capabilities by integrating the `optimx` package,
allowing the use of various optimization algorithms such as "L-BFGS-B", "BFGS", "Nelder-Mead", etc.
}
