% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{P_indiv}
\alias{P_indiv}
\title{Title}
\usage{
P_indiv(p0, est.grid, fix.grid)
}
\arguments{
\item{p0}{}

\item{est.grid}{}

\item{fix.grid}{}
}
\value{

}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
