% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{pd_file}
\alias{pd_file}
\title{Title}
\usage{
pd_file(modelname, .compiledFolder, type = c("indiv", "classic")[1])
}
\arguments{
\item{modelname}{}

\item{.compiledFolder}{}

\item{type}{}
}
\value{

}
\description{
Title
}
