% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{pdIndiv_rebuildPrdObj}
\alias{pdIndiv_rebuildPrdObj}
\title{Rebuild the high-level prediction and objective function}
\usage{
pdIndiv_rebuildPrdObj(pd, Nobjtimes = 100)
}
\arguments{
\item{pd}{pd_indiv}
}
\value{
pd with updated p, prd and obj_data
}
\description{
Rebuild the high-level prediction and objective function
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
