% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{pdIndiv_updateControls}
\alias{pdIndiv_updateControls}
\title{Quickly set "controls" amd update all high-level functions}
\usage{
pdIndiv_updateControls(
  pd,
  rtol = 1e-05,
  atol = 1e-05,
  maxsteps = 5000,
  objtimes = NULL
)
}
\arguments{
\item{pd}{}

\item{rtol}{}

\item{atol}{}

\item{maxsteps}{}

\item{objtimes}{}
}
\value{

}
\description{
Quickly set "controls" amd update all high-level functions
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
