% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{plot.objframe}
\alias{plot.objframe}
\title{Plot method for objframe}
\usage{
\method{plot}{objframe}(x, ...)
}
\arguments{
\item{x}{An objframe object}

\item{...}{Additional arguments passed to plot}
}
\description{
Plot predictions vs observations from an objframe
}
