% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsMalenka.R
\name{plot.parlist}
\alias{plot.parlist}
\title{Plot a parameter list.}
\usage{
\method{plot}{parlist}(x, path = FALSE, ...)
}
\arguments{
\item{x}{fitlist obtained from mstrust}

\item{path}{print path of parameters from initials to convergence. For this
option to be TRUE \code{\link{mstrust}} must have had the option
\option{blather}.}

\item{...}{additional arguments}
}
\description{
Plot a parameter list.
}
\details{
If path=TRUE:
}
\author{
Malenka Mader, \email{Malenka.Mader@fdm.uni-freiburg.de}
}

