% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotArray}
\alias{plotArray}
\title{Plot an array of model predictions for a list of parameters}
\usage{
plotArray(parlist, x, times, data = NULL, ..., fixed = NULL,
  deriv = FALSE, scales = "free", facet = "wrap")
}
\arguments{
\item{parlist}{Matrix or data.frame with a "value" column and columns for the parameters}

\item{x}{The model prediction function \code{x(times, pars, fixed, ...)}}

\item{times}{Numeric vector of time points for the model prediction}

\item{data}{Named list of data.frames as being used in \link{res}, i.e. with columns \code{name}, \code{time},
\code{value} and \code{sigma}.}

\item{...}{Further arguments going to \code{subset}.}

\item{fixed}{Named numeric vector with fixed parameters}

\item{deriv}{Logical. If \code{x} supports the argument \code{deriv}, it is used.}

\item{scales}{The scales argument of \code{facet_wrap} or \code{facet_grid}, i.e. \code{"free"}, \code{"fixed"},
\code{"free_x"} or \code{"free_y"}}

\item{facet}{Either \code{"wrap"} or \code{"grid"}}
}
\value{
A plot object of class \code{ggplot}.
}
\description{
Plot an array of model predictions for a list of parameters
}
\details{
The data.frame being plotted has columns \code{time}, \code{value}, \code{sigma},
\code{name} and \code{condition}.
}

