% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R, R/prdClass.R
\name{plotCombined}
\alias{plotCombined}
\alias{plot.prdlist}
\alias{plotCombined.prdlist}
\alias{plot.prdframe}
\title{Plot a list of model predictions and a list of data points in a combined plot}
\usage{
plotCombined(prediction, ...)

\method{plot}{prdlist}(
  x,
  data = NULL,
  ...,
  scales = "free",
  facet = c("wrap", "grid", "wrap_plain"),
  transform = NULL
)

\method{plotCombined}{prdlist}(
  prediction,
  data = NULL,
  ...,
  scales = "free",
  facet = c("wrap", "grid", "wrap_plain"),
  transform = NULL,
  aesthetics = NULL
)

\method{plot}{prdframe}(
  x,
  data = NULL,
  ...,
  scales = "free",
  facet = c("wrap", "grid", "wrap_plain"),
  transform = NULL
)
}
\arguments{
\item{prediction}{A \code{prdlist} object containing model predictions.}

\item{...}{Filter expressions passed to \code{dplyr::filter} for subsetting
both prediction and data.}

\item{x}{prediction}

\item{data}{Optional data object with observed values. If provided, will be
merged with covariate information and displayed as points with error bars.}

\item{scales}{Scale specification for facets, passed to facet functions.
Default is \code{"free"}.}

\item{facet}{Faceting style. One of:
\itemize{
\item \code{"wrap"}: Facet by name, color by condition (default)
\item \code{"grid"}: Facet grid with name as rows, condition as columns
\item \code{"wrap_plain"}: Facet wrap by name and condition combined
}}

\item{transform}{Optional transformation function applied to coordinates
via \code{coordTransform}.}

\item{aesthetics}{Optional named list of aesthetic mappings (as strings) to
override defaults. Default aesthetics include x, y, ymin, ymax, and
conditionally group and color.}
}
\value{
A plot object of class \code{ggplot}.

A \code{ggplot} object with an additional \code{"data"} attribute
containing a list with the processed \code{data} and \code{prediction}
data frames.
}
\description{
Creates a combined plot of model predictions and observed data with flexible
faceting options. Supports error bars, below limit of quantification (BLoQ)
indicators, and coordinate transformations.
}
\details{
The data.frame being plotted has columns \code{time}, \code{value}, \code{sigma},
\code{name} and \code{condition}.
}
\examples{
\dontrun{    
    ## Observation function
    fn <- eqnvec(
      sine = "1 + sin(6.28*omega*time)",
      cosine = "cos(6.28*omega*time)"
    )
    g <- Y(fn, parameters = "omega")
    
    ## Prediction function for time
    x <- Xt()
    
    ## Parameter transformations to split conditions
    p <- NULL
    for (i in 1:3) {
      p <- p + P(trafo = c(omega = paste0("omega_", i)), condition = paste0("frequency_", i))
    }
    
    ## Evaluate prediction
    times <- seq(0, 1, .01)
    pars <- structure(seq(1, 2, length.out = 3), names = attr(p, "parameters"))
    
    prediction <- (g*x*p)(times, pars)
    
    ## Plotting prediction
    # plot(prediction)
    plotPrediction(prediction)
    plotPrediction(prediction, scales = "fixed")
    plotPrediction(prediction, facet = "grid")
    plotPrediction(prediction, 
                   scales = "fixed",
                   transform = list(sine = "x^2", cosine = "x - 1"))
    
    ## Simulate data
    dataset <- wide2long(prediction)
    dataset <- dataset[seq(1, nrow(dataset), 5),]
    set.seed(1)
    dataset$value <- dataset$value + rnorm(nrow(dataset), 0, .1)
    dataset$sigma <- 0.1
    data <- as.datalist(dataset, split.by = "condition")
    
    ## Plotting data
    # plot(data)
    plot1 <- plotData(data)
     plot1
    ## Plotting data and prediction with subsetting
    # plot(prediction, data)
    plot2 <- plotCombined(prediction, data)
     plot2
    plot3 <- plotCombined(prediction, data, 
                 time <= 0.5 & condition == "frequency_1")
     plot3
    plot4 <- plotCombined(prediction, data, 
                 time <= 0.5 & condition != "frequency_1", 
                 facet = "grid")
     plot4
    plot5 <- plotCombined(prediction, data, aesthetics = list(linetype = "condition"))
     plot5
}
\dontrun{
plotCombined(pred, mydata, time < 100, facet = "grid")
plotCombined(pred, mydata, aesthetics = list(color = "treatment"))
}

}
