% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotFluxes}
\alias{plotFluxes}
\title{Plot Fluxes given a list of flux Equations}
\usage{
plotFluxes(pouter, x, times, fluxEquations, nameFlux = "Fluxes:",
  fixed = NULL)
}
\arguments{
\item{pouter}{parameters}

\item{x}{The model prediction function \code{x(times, pouter, fixed, ...)} needs to return pinner as attribute,
e.g.:\cr
 \code{x <- function(times, pouter, fixed=NULL, attach=TRUE, ...) {\cr
 out <- lapply(conditions, function(cond) { \cr
 pinner <- pL[[cond]](pouter, fixed) \cr
 prediction <- xL[[cond]](times, pinner, ...)\cr
 observation <- g(prediction, pinner, attach.input = attach)\cr
 attr(observation, "pinner") <- pinner\cr
 return(observation)\cr
}); names(out) <- conditions\cr
return(out)}
}}

\item{times}{Numeric vector of time points for the model prediction}

\item{fluxEquations}{list of chars containing expressions for the fluxes,
if names are given, they are shown in the legend. Easy to obtain via \link{subset.eqnList}, see Examples.}

\item{fixed}{Named numeric vector with fixed parameters}
}
\value{
A plot object of class \code{ggplot}.
}
\description{
Plot Fluxes given a list of flux Equations
}
\examples{
\dontrun{
plotFluxes(bestfit,times,attr(subset(f,"B"\%in\%Product), "rates"),nameFlux="B production")
}
}

