% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPaths}
\alias{plotPaths}
\title{Profile likelihood: plot of the parameter paths.}
\usage{
plotPaths(..., whichPar = NULL, sort = FALSE, relative = TRUE,
  scales = "free")
}
\arguments{
\item{...}{Lists of profiles as being returned by \link{profile}.}

\item{whichPar}{Character or index vector, indicating the parameters that are taken as possible reference (x-axis)}

\item{sort}{Logical. If paths from different parameter profiles are plotted together, possible
combinations are either sorted or all combinations are taken as they are.}
}
\value{
A plot object of class \code{ggplot}.
}
\description{
Profile likelihood: plot of the parameter paths.
}

