% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsSvenja.R
\name{plotPathsMulti}
\alias{plotPathsMulti}
\title{Profile likelihood: plot all parameter paths belonging to one profile in one plot}
\usage{
plotPathsMulti(profs, whichpars, npars = 5)
}
\arguments{
\item{profs}{Lists of profiles as being returned by \link{profile}.}

\item{whichpars}{Character vector of parameter names for which the profile paths should be generated.}

\item{npars}{Numeric vector of number of colored and named parameter paths.}
}
\value{
A plot object of class \code{ggplot}.
}
\description{
Profile likelihood: plot all parameter paths belonging to one profile in one plot
}
\examples{
\dontrun{
 plotPathsMulti(myprofiles, c("mypar1", "mypar2"), npars = 5) 
}
}
\author{
Svenja Kemmer, \email{svenja.kemmer@fdm.uni-freiburg.de}
}
