% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{plotPEtabSBML}
\alias{plotPEtabSBML}
\title{Plot observables and states of an SBML/PEtab model}
\usage{
plotPEtabSBML(
  ...,
  g1 = g,
  x1 = x,
  p1 = p0,
  mydata1 = mydata,
  pouter1 = pouter,
  times1 = times,
  errfn = err
)
}
\arguments{
\item{...}{further arguments going to \link{plotCombined}}

\item{g}{observation function as obsfn}

\item{x}{prediction function as prdfn}

\item{p}{parameter function as parfn}

\item{mydata}{data as datalist}

\item{pars}{parameter as vector}

\item{times}{times as vector}
}
\description{
This function plots data and fits of an SBML/PEtab model after the import.
Note: Certain objects generated during the model import by importPEtabSBML have to be present in the global environment and are used as default variables if not specified differntly.
}
\author{
Marcus Rosenblatt and Svenja Kemmer
}
