% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotProfile}
\alias{plotProfile}
\title{Profile likelihood plot}
\usage{
plotProfile(..., maxvalue = 5, parlist = NULL)
}
\arguments{
\item{...}{Lists of profiles as being returned by \link{profile}.}

\item{maxvalue}{Numeric, the value where profiles are cut off.}

\item{parlist}{Matrix or data.frame with columns for the parameters to be added to the plot as points.
If a "value" column is contained, deltas are calculated with respect to lowest chisquare of profiles.}
}
\value{
A plot object of class \code{ggplot}.
}
\description{
Profile likelihood plot
}

