% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotProfilesAndPaths}
\alias{plotProfilesAndPaths}
\title{Profile likelihood: plot profiles along with their parameter paths}
\usage{
plotProfilesAndPaths(
  profs,
  whichpars,
  npars = 5,
  ncols = 3,
  normalizePaths = FALSE,
  modes = c("data", "prior"),
  ...
)
}
\arguments{
\item{profs}{List of profiles as returned by \code{\link[=profile]{profile()}}.}

\item{whichpars}{Character vector of parameter names for which the profile paths should be generated.}

\item{npars}{Numeric indicating number of colored and named parameter paths.}

\item{ncols}{Number of columns in the resulting plot grid.}

\item{normalizePaths}{Logical indicating whether the paths should be normalized to absolute values of 1.
Default \code{FALSE}.}

\item{modes}{Character vector of profile modes to display in the profile plot.
Default \code{c("data", "prior")}. Use e.g. \code{"data"} to show only the data contribution.}

\item{...}{Additional arguments passed to \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}.}
}
\value{
A combined \code{ggplot} object containing the profiles and paths (no shared legend).
}
\description{
Generates combined plots of profile likelihoods and their parameter paths without a shared legend.
}
