% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdIA.R
\name{plotRightVector}
\alias{plotRightVector}
\title{Plot Right Singular Vector from SVDIA Result}
\usage{
plotRightVector(SVDIAResult, index)
}
\arguments{
\item{SVDIAResult}{The result from the SVDIA function.}

\item{index}{The index of the right singular vector to plot.}
}
\value{
A ggplot object showing the right singular vector contributions.
}
\description{
This function plots a specific right singular vector from an SVDIA analysis result.
}
\examples{
result <- SVDIA(prdfun, times)
plotRightVector(result, 1)  # Plot the first right singular vector

}
\author{
Simon Beyer, \email{simon.beyer@fdm.uni-freiburg.de}
}
