% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdIA.R
\name{plotSVDs}
\alias{plotSVDs}
\title{Plot Singular Values from SVDIA Result}
\usage{
plotSVDs(SVDIAResult, sUpper = Inf, sLower = 0)
}
\arguments{
\item{SVDIAResult}{The result from the SVDIA function.}

\item{sUpper}{Upper bound for singular values to plot (default is Inf).}

\item{sLower}{Lower bound for singular values to plot (default is 0).}
}
\value{
A ggplot object showing the singular values.
}
\description{
This function plots the singular values from an SVDIA analysis result,
optionally filtering by upper and lower bounds.
}
\examples{
result <- SVDIA(prdfun, times)
plotSVDs(result)
plotSVDs(result, sUpper = 10, sLower = 1e-3)

}
\author{
Simon Beyer, \email{simon.beyer@fdm.uni-freiburg.de}
}
