% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toolsWolfgang.R
\name{plSelectMin}
\alias{plSelectMin}
\title{Select parameter values with lowest Chi^2 among profiles.}
\usage{
plSelectMin(prf, context = FALSE)
}
\arguments{
\item{prf}{A profile as returned from \link[dMod]{profile}.}

\item{context}{If TRUE, the chi^2 and other context of the profile is
               returned. This parameter defaults to FALSE in which case the
               output can be used as outer parameters directly.}
}
\value{
Parameter values with lowest chi^2 w/ or w/o profile context. If all
        profiles are invalid, NULL is returned.
}
\description{
Select parameter values with lowest Chi^2 among profiles.
}
\details{
On profiling a model, parameter values yielding a lower chi^2 than the one of
the fit providing the optimal values for the profiles might be encountered.
This function extracts the set of parameter values possessing the lowest
chi^2 among all profiles. Profiles which are invalid due to e.g., integration
problem on calculating them, are ignored. If all profiles are invalid, NULL
is returned.
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}

