% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{+.resObjfn}
\alias{+.resObjfn}
\title{Direct sum of residual objective functions}
\usage{
\method{+}{resObjfn}(x1, x2)
}
\arguments{
\item{x1}{Function of class \code{resObjfn}}

\item{x2}{Function of class \code{resObjfn}}
}
\value{
Object of class \code{resObjfn}.
}
\description{
Add two residual objective functions by concatenating their
residuals and Jacobians.
}
\details{
The residual objective functions are evaluated and their residuals
are concatenated. Jacobians are stacked vertically (rbind). Parameters are
matched by name and expanded to include all parameters from both functions.
}
\seealso{
\link{resL2}, \link{constraintResL2}, \link{datapointResL2}
}
