% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{+.resObjlist}
\alias{+.resObjlist}
\title{Add resObj objects}
\usage{
\method{+}{resObjlist}(obj1, obj2)
}
\arguments{
\item{obj1}{resObjlist object}

\item{obj2}{resObjlist object}
}
\value{
Combined resObjlist object
}
\description{
Combine two resObjlist outputs by concatenating residuals and Jacobians.
}
