% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{PRD_indiv}
\alias{PRD_indiv}
\title{Title}
\usage{
PRD_indiv(prd0, est.grid, fix.grid)
}
\arguments{
\item{prd0}{}

\item{est.grid}{}

\item{fix.grid}{}
}
\value{

}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
