% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/prdClass.R
\name{prdlist}
\alias{prdlist}
\alias{as.prdlist}
\alias{as.prdlist.list}
\title{Prediction list}
\usage{
prdlist(...)

as.prdlist(x, ...)

\method{as.prdlist}{list}(x = NULL, names = NULL, ...)
}
\arguments{
\item{...}{objects of class \link{prdframe}
conditions.}

\item{x}{list of prediction frames}

\item{names}{character vector, the list names, e.g. the names of the experimental}
}
\description{
A prediction list is used to store a list of model predictions
from different prediction functions or the same prediction function with different
parameter specifications. Each entry of the list is a \link{prdframe}.
}
