% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDaniel.R
\name{predict.prdfn}
\alias{predict.prdfn}
\title{Model Predictions}
\usage{
\method{predict}{prdfn}(x, times, pars, data = NULL, ...)
}
\arguments{
\item{x}{prediction function}

\item{times}{numeric vector of time points}

\item{pars}{parameter frame, e.g. output from \link{mstrust} or 
\link{profile}}

\item{data}{data list object. If data is passed, its condition.grid
attribute is used to augment the output dataframe by additional 
columns. \code{"data"} itself is returned as an attribute.}

\item{...}{Further arguments goint to the prediction function}
}
\value{
A data frame
}
\description{
Make a model prediction for times and a parameter frame. The
function is a generalization of the standard prediction by a
prediction function object in that it allows to pass a parameter
frame instead of a single parameter vector.
}

