% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDenial.R
\name{predtimes}
\alias{predtimes}
\title{Get a vector of linearly spaced time points for prediction}
\usage{
predtimes(datatimes, eventtimes = NULL, Nobjtimes = 100)
}
\arguments{
\item{datatimes}{}

\item{eventtimes}{}

\item{Nobjtimes}{}
}
\value{

}
\description{
for nice plots
}
