% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqnClass.R
\name{print.eqnvec}
\alias{print.eqnvec}
\title{Print equation vector}
\usage{
\method{print}{eqnvec}(x, width = 140, pander = FALSE, ...)
}
\arguments{
\item{x}{object of class \link{eqnvec}.}

\item{width}{numeric, width of the print-out}

\item{pander}{logical, use pander for output (used with R markdown)}

\item{...}{not used right now}
}
\description{
Print equation vector
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}
