% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsqnl.R
\name{print.lsqnl}
\alias{print.lsqnl}
\title{Print method for lsqnl objects}
\usage{
\method{print}{lsqnl}(x, digits = 6, ...)
}
\arguments{
\item{x}{An object of class \code{lsqnl}}

\item{digits}{Number of significant digits to print (default: 6)}

\item{...}{Additional arguments (ignored)}
}
\value{
The input object (invisibly)
}
\description{
Print method for lsqnl objects
}
