% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme.R
\name{print.nlmeFit}
\alias{print.nlmeFit}
\title{Print an nlmeFit object}
\usage{
\method{print}{nlmeFit}(x, ...)
}
\arguments{
\item{x}{An \code{nlmeFit} object (see \link{nlmeFit}).}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print an nlmeFit object
}
