% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{print.objframe}
\alias{print.objframe}
\title{Print method for objframe}
\usage{
\method{print}{objframe}(x, ...)
}
\arguments{
\item{x}{An objframe object}

\item{...}{Additional arguments passed to print.data.frame}
}
\description{
Print an objframe object
}
