% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmeClass.R
\name{print.omegaSpec}
\alias{print.omegaSpec}
\title{Print method for omega}
\usage{
\method{print}{omegaSpec}(x, ...)
}
\arguments{
\item{x}{An \code{omegaSpec} object.}

\item{...}{Ignored.}
}
\description{
Print method for omega
}
