% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parameters.R
\name{print.par}
\alias{print.par}
\title{Pretty printing of parameter transformations, class par}
\usage{
\method{print}{par}(p, ...)
}
\description{
Pretty printing of parameter transformations, class par
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}

