% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parClass.R
\name{print.parfn}
\alias{print.parfn}
\title{Pretty printing parameter transformations}
\usage{
\method{print}{parfn}(x, ...)
}
\arguments{
\item{x}{prediction function}

\item{...}{additional arguments}
}
\description{
Pretty printing parameter transformations
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}
