% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parClass.R
\name{print.parvec}
\alias{print.parvec}
\title{Pretty printing for parvec objects}
\usage{
\method{print}{parvec}(x, ...)
}
\arguments{
\item{x}{parvec object}

\item{...}{Currently ignored.}
}
\description{
Prints a parameter vector along with information about
its attached derivatives and information about constant parameters in 'fixed'.
}
