% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtabInterface.R
\name{print.petab_problem}
\alias{print.petab_problem}
\title{Print method for \code{petab_problem}}
\usage{
\method{print}{petab_problem}(x, ...)
}
\arguments{
\item{x}{A \code{petab_problem}.}

\item{...}{Unused.}
}
\description{
Print method for \code{petab_problem}
}
