% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtabInterface.R
\name{print.petabProblem}
\alias{print.petabProblem}
\title{Print method for \code{petabProblem}}
\usage{
\method{print}{petabProblem}(x, ...)
}
\arguments{
\item{x}{A \code{petabProblem}.}

\item{...}{Unused.}
}
\description{
Print method for \code{petabProblem}
}
