% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme.R
\name{print.summary.nlmeFit}
\alias{print.summary.nlmeFit}
\title{Print a summary.nlmeFit}
\usage{
\method{print}{summary.nlmeFit}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{summary.nlmeFit} object.}

\item{digits}{Number of significant digits. Default 4.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print a summary.nlmeFit
}
