% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print0}
\alias{print0}
\title{Print object and its "default" attributes only.}
\usage{
print0(x, list_attributes = TRUE)
}
\arguments{
\item{x}{Object to be printed}

\item{list_attributes}{Prints the names of all attribute of x, defaults to
TRUE}
}
\description{
Print object and its "default" attributes only.
}
\details{
Before the \option{x} is printed by print.default, all its arguments
not in the default list of \code{\link[=attrs]{attrs()}} are removed.
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}

Mirjam Fehling-Kaschek,
\email{mirjam.fehling@physik.uni-freiburg.de}
}
