% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsSeverin.R
\name{profile_pars_per_node}
\alias{profile_pars_per_node}
\title{Generate parameter list for distributed profile calculation}
\usage{
profile_pars_per_node(parameters, fits_per_node)
}
\arguments{
\item{parameters}{list of parameters}

\item{fits_per_node}{numerical, number of parameters that will be send to each node.}
}
\value{
List with two arrays: \code{from} contains the number of the starting
parameter, while \code{to} stores the respective upper end of the parameter list
per node.
}
\description{
Generates list of \code{WhichPar} entries to facillitate distribute
profile calculation.
}
\details{
Lists to split the parameters for which the profiles are calculated
on the different nodes.
}
\examples{
\dontrun{
var_list <- profile_pars_node_list(parameter_list, 4)
}

}
