% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/statistics.R
\name{progressBar}
\alias{progressBar}
\title{Progress bar}
\usage{
progressBar(percentage, size = 50, number = TRUE)
}
\arguments{
\item{percentage}{Numeric between 0 and 100}

\item{size}{Integer, the size of the bar print-out}

\item{number}{Logical, Indicates whether the percentage should be printed out.}
}
\description{
Progress bar
}

