% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsWolfgang.R
\name{python.version.rpython}
\alias{python.version.rpython}
\title{Get the Python version to which rPython is linked}
\usage{
python.version.rpython()
}
\value{
The Python version and additional information
}
\description{
Get the Python version to which rPython is linked
}

