% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsWolfgang.R
\name{python.version.sys}
\alias{python.version.sys}
\title{Check which Python versions are installed on the system}
\usage{
python.version.sys(version = NULL)
}
\arguments{
\item{version}{NULL or character. Check for specific version}
}
\value{
Character vector with the python versions and where they are located.
}
\description{
Check which Python versions are installed on the system
}

