% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steadyStates.R
\name{python_version_rpython}
\alias{python_version_rpython}
\title{Get the Python version to which rPython is linked}
\usage{
python_version_rpython()
}
\value{
The Python version and additional information
}
\description{
Get the Python version to which rPython is linked
}
