% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steadyStates.R
\name{python_version_sys}
\alias{python_version_sys}
\title{Check which Python versions are installed on the system}
\usage{
python_version_sys(version = NULL)
}
\arguments{
\item{version}{NULL or character. Check for specific version}
}
\value{
Character vector with the python versions and where they are located.
}
\description{
Check which Python versions are installed on the system
}
