% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRtools.R
\name{read.IQRdata}
\alias{read.IQRdata}
\title{Read IQRsysData file}
\usage{
read.IQRdata(data, regression = NULL)
}
\arguments{
\item{data}{character pointing to the CSV data file}

\item{regression}{names of the regression parameters to be imported}
}
\value{
A datalist object with additional attributes \code{regression} and
\code{dosing} (dosing parameter transformations).
}
\description{
Read IQRsysData file
}
