% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRtools.R
\name{read_IQRdata}
\alias{read_IQRdata}
\title{Read IQRsysData file}
\usage{
read_IQRdata(data, keep = NULL, split.by = "CONDITION", loq = -Inf)
}
\arguments{
\item{data}{character pointing to the CSV data file}

\item{keep}{names of the elements to keep, e.g. regression parameters to be imported}

\item{split.by}{column by which to split into conditions. If other than "CONDITION" must be contained in `keep`.}

\item{loq}{limit of quantification, single numeric or named numerig, e.g. `c(OUTPUT1 = -3, OUTPUT2 = -5)`.}
}
\value{
A datalist object with additional attributes \code{keep} and
\code{dosing} (dosing parameter transformations).
}
\description{
Read IQRsysData file
}
