% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{reduceReplicates.character}
\alias{reduceReplicates.character}
\title{Method for files (character)}
\usage{
\method{reduceReplicates}{character}(data, select = "condition", datatrans = NULL)
}
\description{
Method for files (character)
}
