% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{reduceReplicates.data.frame}
\alias{reduceReplicates.data.frame}
\title{Method for data frames}
\usage{
\method{reduceReplicates}{data.frame}(data, select = "condition", datatrans = NULL)
}
\description{
Method for data frames
}
