% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{rename_objlist}
\alias{rename_objlist}
\title{Rename the derivatives of an objective function according to a lookup table}
\usage{
rename_objlist(myobjlist, condition, est.grid)
}
\arguments{
\item{myobjlist}{dMod objective list}

\item{condition}{a specific condition (ID)}

\item{est.grid}{estimation grid}
}
\value{
objective list
}
\description{
Rename the derivatives of an objective function according to a lookup table
}
\author{
Daniel Lill
}
