% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{renameDerivPars}
\alias{renameDerivPars}
\title{Title}
\usage{
renameDerivPars(pred0, pars, est.grid, cn)
}
\arguments{
\item{pred0}{prd0(times,pars)[[1]]}

\item{pars}{pars}

\item{est.grid}{est.grid}

\item{cn}{name of condition}
}
\value{
pred0 with updated deriv argument
}
\description{
Title
}
