% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{renameDerivParsInObjlist}
\alias{renameDerivParsInObjlist}
\title{Rename the gradient and hessian of an objlist}
\usage{
renameDerivParsInObjlist(objlist, parnames)
}
\arguments{
\item{objlist}{an objlist with grid.inner parnames}

\item{parnames}{vector(grid.inner = grid.outer), e.g. as the \code{parnames} element from the result of \code{\link[=make_pars]{make_pars()}}}
}
\value{
objlist with new names
}
\description{
Rename the gradient and hessian of an objlist
}
\examples{
ol <- dMod:::init_empty_objlist(c(a = 1, b = 2))
parnames <- c(a = "c", b = "d")
renameDerivParsInObjlist(ol, parnames)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
