% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repWithNames}
\alias{repWithNames}
\title{Named repititions}
\usage{
repWithNames(x, names)
}
\arguments{
\item{x}{Value to be repeated.}

\item{names}{List of names.}
}
\description{
Wrapper on rep() to input names instead of length.
}
