% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{res}
\alias{res}
\title{Compare data and model prediction by computing residuals}
\usage{
res(data, out, err = NULL)
}
\arguments{
\item{data}{data.frame with columns: time, name, value, sigma, lloq}

\item{out}{matrix with predictions (first column = time), optional "deriv" attribute}

\item{err}{optional matrix with error model predictions, optional "deriv" attribute}
}
\value{
objframe with attributes "deriv", "deriv.err"
}
\description{
Compare data and model prediction by computing residuals
}
