% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{res_cpp}
\alias{res_cpp}
\title{Compute residuals (Rcpp implementation)}
\usage{
res_cpp(data, out, err = NULL)
}
\arguments{
\item{data}{DataFrame with columns: time, name, value, sigma, lloq}

\item{out}{NumericMatrix with predictions (first column = time)}

\item{err}{Nullable NumericMatrix with error model predictions}
}
\value{
List with data and derivatives (3D/4D arrays)
}
\description{
Compute residuals (Rcpp implementation)
}
