% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolics.R
\name{rref}
\alias{rref}
\title{Transform a matrix into reduced row echelon form}
\usage{
rref(A, tol = sqrt(.Machine$double.eps), verbose = FALSE, fractions = FALSE)
}
\arguments{
\item{A}{Numeric matrix for which the reduced row echelon form is computed.}

\item{tol}{Numeric tolerance used to identify pivot elements. Defaults to \code{sqrt(.Machine$double.eps)}.}

\item{verbose}{Logical. If \code{TRUE}, prints detailed information during computation.}

\item{fractions}{Logical. Currently not used.}
}
\value{
A list with two elements:
\itemize{
\item \code{[[1]]}: The reduced row echelon form of \code{A}.
\item \code{[[2]]}: The indices of the columns in which pivots were found.
}
}
\description{
This function computes the reduced row echelon form (RREF) of a numeric matrix.
It is written along the lines of the MATLAB function \code{rref}.
}
\author{
Malenka Mader, \email{Malenka.Mader@fdm.uni-freiburg.de}
}
