% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoteComputing.R
\name{runbg.install}
\alias{runbg.install}
\title{Remote install dMod to a ssh-reachable host}
\usage{
runbg.install(type = "local", sshtarget, source = NULL)
}
\arguments{
\item{type}{Which dMod to install. At the moment, only your local version is
supported.}

\item{sshtarget}{The ssh host url.}

\item{source}{If type = local, source must point to the source directory of
your dMod version. This is most probably you local dMod git repository.}
}
\description{
Install your local dMod version to a remote host via ssh.
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}

