% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{savedMod}
\alias{savedMod}
\title{Save snapeshot of your dMod project}
\usage{
savedMod(..., objfn = NULL, files = NULL,
  filename = stop("'filename' must be specified"))
}
\arguments{
\item{...}{objects in the global environment (as symbols or 
character strings) to be saved \strong{in addition to} standard
dMod objects as being return by \link{lsdMod} and data list objects.}

\item{objfn}{character specifying the objective function, e.g. 
\code{"normL2(data, g*x*p) + constr"}.}

\item{files}{additional files to be saved, e.g. script files.}

\item{filename}{name of the file where everything is saved.}
}
\value{
A zip file is written in the working directory containing
an image file with the dMod objects, C files associated to the
dMod objects and additional files specified by the 'files' argument.
}
\description{
Save snapeshot of your dMod project
}
\examples{
\dontrun{
savedMod(pouter, bestfit, fitlist,
         objfn = "normL2(data, g*x*p) + constr",
         files = c("modelsetup.R", "analysis.R"),
         filename = "snapshot_161207")

} 
}
\seealso{
\link{loaddMod}
}
