% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsMirjam.R
\name{saveShiny}
\alias{saveShiny}
\title{Save the necessary objects for dModtoShiny with the correct names}
\usage{
saveShiny(
  reactions = myeqnlist,
  x,
  parameters,
  fixed = NULL,
  data,
  profiles,
  pubref = "none",
  errmodel = NULL
)
}
\arguments{
\item{reactions}{eqnlist-object}

\item{x}{prediction funciton}

\item{parameters}{Parframe, result of mstrust()}

\item{fixed}{Character? fixed parameters?}

\item{data}{datalist}

\item{profiles}{Parframe, result of profile(). Is a proflist also possible?}

\item{pubref}{Link to publication}

\item{errmodel}{obsfn}
}
\description{
Save the necessary objects for dModtoShiny with the correct names
}
