% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDenial.R
\name{saveShiny_dMod.frame}
\alias{saveShiny_dMod.frame}
\title{SaveShiny for dMod.frames}
\usage{
saveShiny_dMod.frame(dMod.frame, hypothesis = 1,
  reactions = dMod.frame$reactions[[hypothesis]], pubref = "none",
  fixed = dMod.frame$fixed[[hypothesis]], projectname)
}
\arguments{
\item{dMod.frame}{dmod.frame}

\item{hypothesis}{hypothesis}

\item{reactions}{eqnlist which are the basis for prd}

\item{pubref}{character. pubmed link or similar}

\item{fixed}{}

\item{projectname}{Name of the folder being created on fermi}
}
\description{
SaveShiny for dMod.frames
}
