% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scale_fill_dMod}
\alias{scale_fill_dMod}
\title{Standard dMod fill scheme}
\usage{
scale_fill_dMod(...)
}
\arguments{
\item{...}{arguments forwarded to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
Standard dMod fill scheme
}
