% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsPaul.R
\name{scaledependent}
\alias{scaledependent}
\title{Scale dependent flux sensitivity Analysis}
\usage{
scaledependent(
  frame,
  x,
  obstates,
  bestfit,
  time,
  p,
  bound = 0.001,
  figurename = "scaledependent",
  timeunit = "h",
  width = 16,
  height = 9,
  linewidth = 2
)
}
\arguments{
\item{frame}{Reactionlist of the model}

\item{x}{Xs function of the model}

\item{obstates}{The states under observation whose changes are analysed}

\item{bestfit}{The parameter values which are used for the simulation}

\item{time}{Array of observed timepoints}

\item{p}{P function of the system}

\item{bound}{Absolute value which a flux sensitive component has to reach at any point in time to be shown in the outputplots}

\item{figurename}{Name of the saved pdf}

\item{timeunit}{Unit of the time at the x-axis of the plots}

\item{linewidth}{Linewidth of the coloured fluxes in the plots}

\item{time, width}{Dimension of the plots}
}
\description{
This function performs the scale dependent flux sensitivity analysis for a simulated system. It prints a pdf-file to the working directory for each condition of the system.
}
\examples{
#Scale dependent flux sensitivity Analysis for the SIR-Model with two different values for beta

frame <- NULL
frame <-addReaction(frame, from = "S", to = "I", rate = "beta*S*I")
frame <-addReaction(frame, from = "I", to = "R", rate = "gamma*I")
x <-Xs(odemodel(frame))

obstates = c("S", "I", "R")

par <- getParameters(x)
p <- NULL
for (i in c(0.0026, 0.0027)){
  trafo <- repar("x~x", x=par)
  trafo <- repar("x~y", x = "beta", y= i, trafo)
  parameter <- structure(trafo, names = par)
  p <- p+P(parameter, condition= paste0(i))}


time <- seq(0, 14, length.out=1000)
bestfit <- c(S = 7.637603e+02, I = 6.184001e-01, R = 8.251977e-16, gamma = 4.582934e-01)

scaledependent(frame, x, obstates, bestfit, time, p)
}
