% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R2CdeSolve.R
\name{setForcings.einspline}
\alias{setForcings.einspline}
\title{Initialize the forcings, i.e. fill the splines used as forcing functions in the C model file}
\usage{
\method{setForcings}{einspline}(func, forcings)
}
\arguments{
\item{func}{The return value of \link{funC.einspline}}

\item{forcings}{A \code{data.frame} with columns "name", "time" and "value" specifying the
forcing function.}
}
\value{
A list of matrices with times and values, one matrix for each forcing.
}
\description{
Initialize the forcings, i.e. fill the splines used as forcing functions in the C model file
}
\details{
Running \code{setForcings()}, the C function \code{createSplines()} is called via
the \code{.C()} interface. In addition, the data frame is translated into a list of matrices
as being encountered when using the forcings interface of the \code{deSolve} package.
}

