% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R2CdeSolve.R
\name{setForcings}
\alias{setForcings}
\title{Initialize the forcings, i.e. fill the splines used as forcing functions in the C model file}
\usage{
setForcings(func, forcings)
}
\description{
Initialize the forcings, i.e. fill the splines used as forcing functions in the C model file
}
\details{
See \link{setForcings.einspline} and \link{setForcings.nospline}.
}

