% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsSeverin.R
\name{steadyStateToolJulia}
\alias{steadyStateToolJulia}
\title{Calculated the steady states of a given model}
\usage{
steadyStateToolJulia(el, forcings = NULL, neglect = NULL, verboseLevel = 1)
}
\arguments{
\item{el}{the equation list}

\item{forcings}{vector of strings, default \code{c("","")}. The names of the forcings which will be set to zero.}

\item{neglect}{vector of strings, default \code{c("","")}. The names of the variables which will be neglected as fluxParameters and therefore will not be solved for.}

\item{verboseLevel}{integer, default \code{1}. The level of verbosity of the output, right now only 1 (all) and 0 (no) is implementes.}
}
\value{
named vector with the steady state transformations. The names are the inner, the values are the outer parameters
}
\description{
Uses julia to calculate the steady state transformations
}
