% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{strelide}
\alias{strelide}
\title{Elide character vector}
\usage{
strelide(string, width, where = "right", force = FALSE)
}
\arguments{
\item{string}{String subject to eliding}

\item{width}{Width including eliding ... of return string}

\item{where}{Eliding can happen at 'left', 'middel', or 'right'. Defaults to
'right'.}

\item{force}{Elide, even is \if{html}{\out{<string>}} is shorter than \if{html}{\out{<width>}}. Default to
'FALSE'.}
}
\value{
Elided string of length \if{html}{\out{<width>}}.
}
\description{
Elide character vector
}
\details{
Elide a string to \if{html}{\out{<width>}}. Eliding can happen at 'left', 'middle', or
'right'. #' If forcing = FALSE, which is the default, strings shorten than
\if{html}{\out{<width>}} are returend unaltered; forcing = TRUE inserts eliding symbols (...)
in any case.
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}
