% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{sumDuplicatedParsInDeriv}
\alias{sumDuplicatedParsInDeriv}
\title{Sum redundant columns}
\usage{
sumDuplicatedParsInDeriv(der)
}
\arguments{
\item{der}{deriv of a single condition, as used in [renameDerivPars]}
}
\value{
der with redundant columns summed and duplicates removed
}
\description{
Remove redundancies (happens when a parameter is duplicated and mapped to the same outer parameter). 
For example in est.grid = data.table(ID = 1, init_A = "Atot", Atot= "Atot", ...)
}
