% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{sumDuplicatedParsInObjlist}
\alias{sumDuplicatedParsInObjlist}
\title{Remove redundant outer names}
\usage{
sumDuplicatedParsInObjlist(ol)
}
\arguments{
\item{objlist}{objlist with potentially duplicated names}
}
\value{
objlist with duplicated gradient and hessian elements summed and redundancies removed
}
\description{
Remove redundancies (happens when a parameter is duplicated and mapped to the same outer parameter).
For example in est.grid = data.table(ID = 1, init_A = "Atot", Atot= "Atot", ...)
}
\examples{
ol <- dMod:::init_empty_objlist(c("S2" = 2, "S3" = 3, S2 = 3))
ol$gradient <- ol$gradient + 1:3
ol$hessian <- ol$hessian + 1:9
sumDuplicatedParsInObjlist(ol)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
