% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqnClass.R
\name{summary.eqnvec}
\alias{summary.eqnvec}
\title{Summary of an equation vector}
\usage{
\method{summary}{eqnvec}(object, ...)
}
\arguments{
\item{object}{of class \link{eqnvec}.}

\item{...}{additional arguments}
}
\description{
Summary of an equation vector
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}
