% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsqnl.R
\name{summary.lsqnl}
\alias{summary.lsqnl}
\title{Summary method for lsqnl objects}
\usage{
\method{summary}{lsqnl}(object, ...)
}
\arguments{
\item{object}{An object of class \code{lsqnl}}

\item{...}{Additional arguments (ignored)}
}
\value{
Object of class \code{summary.lsqnl} with additional summary statistics
}
\description{
Provides detailed summary statistics for the optimization result.
}
