% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{summary.objframe}
\alias{summary.objframe}
\title{Summary method for objframe}
\usage{
\method{summary}{objframe}(object, ...)
}
\arguments{
\item{object}{An objframe object}

\item{...}{Additional arguments (not used)}
}
\description{
Summarize an objframe object
}
