% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdIA.R
\name{SVDIA}
\alias{SVDIA}
\title{SVD-based Identifiability Analysis}
\usage{
SVDIA(
  prdfun,
  times,
  HC = list(parlower = -5, parupper = 3),
  nsamples = 10,
  neglect = NULL,
  cores = 1
)
}
\arguments{
\item{prdfun}{The model function for which the sensitivity matrix is calculated.}

\item{times}{Vector of time points for evaluation.}

\item{HC}{A list with the parameter bounds. The list can be in two forms:
- `HC = list(parlower = lower_value, parupper = upper_value)` for the same bounds for all parameters.
- `HC = list(parlower = named_numeric_vector, parupper = named_numeric_vector)` for specific bounds for each parameter.
Defaults to `list(parlower = -5, parupper = 3)`.}

\item{nsamples}{The number of random samples to generate (default is 10).}

\item{neglect}{Vector of parameter names to exclude from the analysis (default is NULL).}

\item{cores}{Number of cores to use for parallel computation (default is 1).}
}
\value{
A list containing:
  - `tilde_S`: The concatenated sensitivity matrix.
  - `svd_result`: The result of the Singular Value Decomposition.
  - `samples`: The randomly sampled parameter sets.
  - `singular_values`: The singular values from SVD.
  - `right_singular_vectors`: The right singular vectors from SVD.
  - `param_names`: Names of parameters included in the analysis.
}
\description{
This function performs a Singular Value Decomposition (SVD) based identifiability analysis
using Sobol sampling to explore parameter sets for a given prediction function (usually 'g*x*p').
It calculates the sensitivity matrix for each set of parameters, performs SVD,
and returns the numerical results.
}
\examples{
# Assuming you have a model function `prdfun` and parameter bounds `HC`
HC <- list(parlower = -5, parupper = 5)  # Default bounds for all parameters
result <- SVDIA(prdfun, times, HC, nsamples = 100)
plotSVDs(result)
plotRightVector(result, 1)  # Plot the first right singular vector

}
\author{
Simon Beyer, \email{simon.beyer@fdm.uni-freiburg.de}
}
