% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{testPEtabSBML}
\alias{testPEtabSBML}
\title{Test PEtabSBML import}
\usage{
testPEtabSBML(
  models = c("0001", "0002", "0003", "0004", "0005", "0006", "0007", "0008", "0009",
    "0010", "0011", "0012", "0013", "0014", "0015", "0016"),
  testFit = TRUE,
  timelimit = 5000,
  testCases = TRUE
)
}
\arguments{
\item{models}{model names to test}
}
\value{
evaluation data.frame
}
\description{
This function imports, evaluates and tests the PEtab model.
}
\author{
Marcus Rosenblatt and Svenja Kemmer
}
