% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_dMod}
\alias{theme_dMod}
\title{Standard plotting theme of dMod}
\usage{
theme_dMod(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{numeric, font-size}

\item{base_family}{character, font-name}
}
\description{
Standard plotting theme of dMod
}

