% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRtools.R
\name{to_dMod.frame}
\alias{to_dMod.frame}
\title{Export to IQRtools imports to dMod.frame format}
\usage{
to_dMod.frame(hypothesis = date(), data, model, errmodel = NULL,
  transform = NULL)
}
\arguments{
\item{hypothesis}{hypothesis name}

\item{data}{result from \link{read.IQRdata()}}

\item{model}{result from \link{read.IQRmodel()}}

\item{errmodel}{an error model}

\item{transform}{vector with transformations ("L" (log), "N" (normal) or "G" (log-it))}
}
\value{
A \linke{dMod.frame} object
}
\description{
Export to IQRtools imports to dMod.frame format
}
