% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parClass.R
\name{unique.parframe}
\alias{unique.parframe}
\title{Extract those lines of a parameter frame with unique elements in the value column}
\usage{
\method{unique}{parframe}(x, incomparables = FALSE, tol = 1, ...)
}
\arguments{
\item{x}{parameter frame}

\item{incomparables}{not used. Argument exists for compatibility with S3 generic.}

\item{tol}{tolerance to decide when values are assumed to be equal, see \code{\link{plotValues}()}.}

\item{...}{additional arguments being passed to \code{\link{plotValues}()}, e.g. for subsetting.}
}
\value{
A subset of the parameter frame \code{x}.
}
\description{
Extract those lines of a parameter frame with unique elements in the value column
}
