% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{vcov}
\alias{vcov}
\title{Get variance-covariance matrix from trust result}
\usage{
vcov(fit, parupper = NULL, parlower = NULL)
}
\arguments{
\item{fit}{return from trust or selected fit from parlist}

\item{parupper}{upper limit for parameter values. Parameters
are considered fixed when exceeding these limits.}

\item{parlower}{lower limit for parameter values. Parameters
are considered fixed when exceeding these limits.}
}
\description{
Get variance-covariance matrix from trust result
}
