% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{wide2long.list}
\alias{wide2long.list}
\title{Translate wide output format (e.g. from ode) into long format}
\usage{
\method{wide2long}{list}(out, keep = 1, na.rm = FALSE)
}
\arguments{
\item{out}{list of matrices in wide format}

\item{keep}{Index vector, the columns to keep}

\item{na.rm}{Logical, if \code{TRUE}, missing values are removed in the long format.}
}
\value{
data.frame in long format, i.e. columns "time" (out[,1]), "name" (colnames(out[,-1])),
"value" (out[,-1]) and, if out was a list, "condition" (names(out))
}
\description{
Translate wide output format (e.g. from ode) into long format
}
\details{
The function assumes that out[,1] represents a time-like vector
whereas out[,-1] represents the values. Useful for plotting with ggplot. If
a list is supplied, the names of the list are added as extra column names "condition"
}

