% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqnClass.R
\name{write.eqnlist}
\alias{write.eqnlist}
\title{Write equation list into a csv file}
\usage{
write.eqnlist(eqnlist, ...)
}
\arguments{
\item{eqnlist}{object of class \link{eqnlist}}

\item{...}{Arguments going to \link[utils]{write.table}}
}
\description{
Write equation list into a csv file
}
