% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod2PEtab_Svenja.R
\name{writeMeasurementsTSV}
\alias{writeMeasurementsTSV}
\title{Export Measurement Data as .tsv}
\usage{
writeMeasurementsTSV(mymodel)
}
\arguments{
\item{data}{PEtab data file as .tsv}

\item{observables}{observables as eqnvec}
}
\value{
data as data list and errors (if required) as eqnvec.
}
\description{
This function imports data from the PEtab data file as a data list and defines errors if an error model is required.
}
\author{
Marcus Rosenblatt and Svenja Kemmer
}
