% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod2PEtab_Svenja.R
\name{writeObservablesTSV}
\alias{writeObservablesTSV}
\title{Export observables as .tsv}
\usage{
writeObservablesTSV(observables, errors, modelname)
}
\arguments{
\item{observables}{dMod observable as named vector or eqnvec}
}
\value{
PEtab observable file as .tsv
}
\description{
This function exports observables from dMod into the PEtab observable file.
}
\author{
Svenja Kemmer
}
