% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqnClass.R
\name{x2logx}
\alias{x2logx}
\title{Log-transform variables in an equation vector using SymPy}
\usage{
x2logx(eqnvec, whichVar)
}
\arguments{
\item{eqnvec}{An object of class \code{eqnvec}, representing a set of equations.}

\item{whichVar}{A character vector specifying the variables to be log-transformed.}
}
\value{
An updated \code{eqnvec} object with transformed equations.
}
\description{
Log-transform variables in an equation vector using SymPy
}
\details{
Replaces each \code{var} in \code{whichVar} with \code{exp(log_var)} and rewrites
the corresponding ODE for \code{log_var = log(var)}. Output equations are
simplified via SymPy.
}
\examples{
eqnvec <- as.eqnvec(
  c("-k1*A + k2*B", "k1*A - k2*B"),
  names = c("A", "B")
)
log_transformed_eqns <- x2logx(eqnvec, c("A", "B"))

}
