% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prediction.R
\name{Xv}
\alias{Xv}
\title{Model evaluation.}
\usage{
Xv(func_fa, func_l, forcings = NULL, events = NULL, data,
  optionsBvp = NULL, optionsOde = list(method = "lsode"),
  optionsData = list())
}
\arguments{
\item{func_fa}{Return value of \code{funC(fa)} where \code{fa} defines the ODE and the
adjoint sensitivities. (See also \link{generateModelIE}).}

\item{func_l}{Return value of \code{funC(l))} where \code{l} defines the ODE for the
time-continuous chi² function and its gradient. (See also \link{generateModelIE}).}

\item{forcings}{A data.frame with columns name (factor), time (numeric) and value (numeric).
The ODE forcings.}

\item{events}{data.frame of events (Not yet functional in bvptwp).}

\item{data}{A data.frame with columns "name", "time", "value" and "sigma".}

\item{optionsBvp}{list with arguments to be passed to \link{bvptwpC} for solving the BVP problem func_fa.}

\item{optionsOde}{list with arguments to be passed to \link{odeC} for the ODE integration of func_l.}

\item{optionsData}{list with arguments to be passed to \link{data2forc} for generation of
the data representation functions.}
}
\value{
a function \code{x(times, pars, forcings, events, eps, atol, nmax, guess, deriv)} representing the model evaluation.
The default values are \code{eps = 1}, \code{atol = 1e-4}, \code{nmax = 50*length(times)}, \code{guess = NULL}
and \code{deriv = TRUE}.
The result of
\code{x(times, pars, ...)} has attributes "value" (the chi² value)
and "grad" (gradient of the chi² value) when \code{deriv=TRUE}.
If \code{deriv=FALSE}, the time-continuous chi² is not evaluated.
If \code{pars} is
the result of \code{p(P)} (see \link{P}), the Jacobian of the parameter transformation
and the gradient of the chi² are multiplied according to the chain rule for
differentiation.
}
\description{
Interface to combine an ODE, its adjoint sensitivity equations as well
as the chi² and the gradient of chi²
into one model function \code{x(times, pars, ...)}.
}

