% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{check_cond}
\alias{check_cond}
\title{Do some consistency checks on conditional and condition.grid}
\usage{
check_cond(conditional, condition.grid)
}
\arguments{
\item{conditional}{see \link{normL2}}

\item{condition.grid}{condition.grid from data}
}
\description{
Do some consistency checks on conditional and condition.grid
}
