% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{check_grids}
\alias{check_grids}
\title{Some consistency checks for fixed.grid and est.grid}
\usage{
check_grids(fixed.grid, est.grid)
}
\arguments{
\item{fixed.grid, est.grid}{data.frame(parname, partask, ids...)}
}
\value{
TRUE: All tests passed, else an error is thrown
}
\description{
* Checks for localized parameters appearing in both grids that exactly one NA is in either of the grids
* More to come ...
}
\examples{
fg <- stats::setNames(data.frame(parname = "d", partask = "Cond_specific", "1" = 1, "2" = NA, "3" = NA, stringsAsFactors = FALSE), c("parname", "partask", as.character(1:3)))
eg_good <- stats::setNames(data.frame(parname = "d", partask = "Cond_specific", "1" = NA, "2" = "dummy", "3" = "dummy", stringsAsFactors = FALSE), c("parname", "partask", as.character(1:3)))
eg_bad1 <- stats::setNames(data.frame(parname = "d", partask = "Cond_specific", "1" = NA, "2" = NA, "3" = "dummy", stringsAsFactors = FALSE), c("parname", "partask", as.character(1:3)))
eg_bad2 <- stats::setNames(data.frame(parname = "d", partask = "Cond_specific", "1" = "dummy", "2" = "dummy", "3" = "dummy", stringsAsFactors = FALSE), c("parname", "partask", as.character(1:3)))

check_grids(fg, eg_good)
check_grids(fg, eg_bad1)
check_grids(fg, eg_bad2)
}
