% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{compile}
\alias{compile}
\title{Compile one or more prdfn, obsfn or parfn objects}
\usage{
compile(..., output = NULL, args = NULL, cores = 1, verbose = F)
}
\arguments{
\item{...}{Objects of class parfn, obsfn or prdfn}

\item{output}{Optional character of the file to be produced. If several objects were
passed, the different C files are all compiled into one shared object file.}

\item{args}{Additional arguments for the R CMD SHLIB call, e.g. \code{-leinspline}.}

\item{cores}{Number of cores used for compilation when several files are compiled.}

\item{verbose}{Print compiler output to R command line.}
}
\description{
Compile one or more prdfn, obsfn or parfn objects
}
