% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataClass.R
\name{covariates}
\alias{covariates}
\title{Access the covariates in the data}
\usage{
covariates(data)
}
\arguments{
\item{data}{Either a \link{datalist} or a \code{data.frame} with mandatory columns \code{c("name", "time", "value", "sigma")}}
}
\value{
The \code{condition.grid}
}
\description{
Access the covariates in the data
}
