% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{datapointL2_indiv}
\alias{datapointL2_indiv}
\title{DatapointL2 without access to stored predictions}
\usage{
datapointL2_indiv(
  name,
  time,
  value,
  sigma = 1,
  attr.name = "validation",
  condition,
  prd_indiv
)
}
\arguments{
\item{name}{}

\item{time}{}

\item{value}{}

\item{sigma}{}

\item{attr.name}{}

\item{condition}{}

\item{prd_indiv}{}
}
\value{

}
\description{
DatapointL2 without access to stored predictions
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
