% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoteComputing.R
\name{detectFreeCores}
\alias{detectFreeCores}
\title{Detect number of free cores (on UNIX)}
\usage{
detectFreeCores(machine = NULL)
}
\arguments{
\item{machine}{character, e.g. "user@localhost".}
}
\description{
Read \code{/proc/loadavg} and subtract from the number of cores
}

