% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModFrameClass.R
\name{dMod.frame}
\alias{dMod.frame}
\title{Generate a dMod.frame}
\usage{
dMod.frame(hypothesis, g, x, p, data, e = NULL, ...)
}
\arguments{
\item{hypothesis}{Character. Description of the hypothesis}

\item{g}{fn}

\item{x}{fn}

\item{p}{fn}

\item{data}{data.frame or datalist, will be coerced to datalist}

\item{e}{fn}

\item{...}{other columns, have to be named. They will be placed in a list.}
}
\value{
Object of class \code{tbl_df}, which is grouped rowwise.
}
\description{
Basically, a dMod.frame is a \link{tibble}, which is grouped \link{rowwise}.

Since the dMod.frame is also designed for interactive use, the class will not be called
"dMod.frame" as I initially planned, but will be c("tbl_df").
This way, I don't have to overwrite all the dplyr-verbs.

The dMod.frame object stores all objects that are needed to reproducibly
produce a result, such as a plot or profiles, in one row of a \link{tibble}.
Each row corresponds to a distinct hypothesis, e.g. one would have two distinct rows
for fitting with and without a prior.
}
