% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toolsWolfgang.R
\name{fitErrorModel}
\alias{fitErrorModel}
\title{Fit an error model}
\usage{
fitErrorModel(data, factors, errorModel = "exp(s0)+exp(srel)*x^2",
  par = c(s0 = 1, srel = 0.1), plotting = TRUE, ...)
}
\arguments{
\item{data}{Reduced replicate data, see \code{\link{reduceData}}.}

\item{factors}{\option{data} is pooled with respect to the columns named
here, see Details.}

\item{errorModel}{Character vector defining the error model. Use \kbd{x} to
reference the independend variable, see Details.}

\item{par}{Inital values for the parameters of the error model.}

\item{plotting}{If TRUE, a plot of the pooled variance together with the fit
of the error model is shown.}

\item{...}{Parameters handed to the optimizer \code{\link{optim}}.}
}
\value{
Returned is a data frame with the first columns identical to
  \option{data}. Appended are fit values of the parameters of the error
  model, with the column names equal to the parameter names and the estimated
  uncertainty Sigma. Sigma is derived by evaluating the error model with the
  fit parameters. The error model is appended as the attribute "errorModel".

  Confidence bounds for Sigma at confidence level 68\% and 95\% are
  calculated, Their values come next in the returned data frame. Finally, the
  effective conditions are appended to easily check how the pooling was done.
}
\description{
Fit an error model to reduced replicate data, see
  \code{\link{reduceReplicates}}.
}
\details{
The variance estimator using \eqn{n-1} data points is \eqn{chi^2}
  distributed with \eqn{n-1} degrees of freedom. Given replicates for
  consecutive time points, the sample variance can be assumed a function of
  the sample mean. By defining an error model which must hold for all time
  points, a maximum likelihood estimator for the parameters of the error
  model can be derived. The parameter \option{errorModel} takes the error
  model as a character vector, where the mean (independent variable) is
  refered to as \kbd{x}.

  It is desireable to estimate the variance from many replicates. The
  parameter \option{data} must provide one or more columns which define the
  pooling of data. In case more than one column is announced by
  \option{factors}, all combinations are constructed. If, e.g.,
  \option{factors = c("Condition", "Species")} is used, where "Condition" is
  "a", "b", "c" and repeating and Species is "d", "e" and repeating, the
  effective conditions used for pooling are "a d", "b e", "c d", "a e", "b
  d", and "c e".

  By default, a plot of the pooled data, Sigma and its confidence bound at
  68\% and 95\% is shown.
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}

