% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolics.R
\name{forcingsSymb}
\alias{forcingsSymb}
\title{Return some useful forcing functions as strings}
\usage{
forcingsSymb(
  type = c("Gauss", "Fermi", "1-Fermi", "MM", "Signal", "Dose"),
  parameters = NULL
)
}
\arguments{
\item{type}{Which function to be returned}

\item{parameters}{Named vector, character or numeric. Replace parameters by the corresponding valus
in \code{parameters}.}
}
\value{
String with the function
}
\description{
Return some useful forcing functions as strings
}
