% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{getDerivs}
\alias{getDerivs}
\alias{getDerivs.list}
\alias{getDerivs.objlist}
\alias{getDerivs.parvec}
\alias{getDerivs.prdframe}
\alias{getDerivs.prdlist}
\title{Extract the derivatives of an object}
\usage{
getDerivs(x, ...)

\method{getDerivs}{parvec}(x, ...)

\method{getDerivs}{prdframe}(x, ...)

\method{getDerivs}{prdlist}(x, ...)

\method{getDerivs}{list}(x, ...)

\method{getDerivs}{objlist}(x, ...)
}
\arguments{
\item{x}{object from which the derivatives should be extracted}

\item{...}{additional arguments (not used right now)}
}
\value{
The derivatives in a format that depends on the class of \code{x}.
This is
\code{parvec -> matrix}, 
\code{prdframe -> prdframe}, 
\code{prdlist -> prdlist}, 
\code{objlist -> named numeric}.
}
\description{
Extract the derivatives of an object
}

