% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{getDerivs2}
\alias{getDerivs2}
\alias{getDerivs2.parvec}
\alias{getDerivs2.prdframe}
\title{Extract the second derivatives of an object}
\usage{
getDerivs2(x, full = FALSE, ...)

\method{getDerivs2}{parvec}(x, ...)

\method{getDerivs2}{prdframe}(x, full = FALSE, ...)
}
\arguments{
\item{x}{Object from which the second derivatives should be extracted.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The second derivatives, in a format depending on the class of \code{x}.
For now:
\itemize{
  \item \code{parvec -> 3D array (Hessian tensor)}
}
}
\description{
Generic function to extract second derivatives (Hessians)
from various model objects.
}
