% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{getObservables}
\alias{getObservables}
\title{Extract the observables of an object}
\usage{
getObservables(x, ...)
}
\arguments{
\item{x}{object from which the equations should be extracted}

\item{...}{not used}
}
\value{
The equations as a character.
}
\description{
Extract the observables of an object
}
