% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{getParameters}
\alias{getParameters}
\alias{getParameters_dispatch}
\alias{getParameters_dispatch.odemodel}
\alias{getParameters_dispatch.fn}
\alias{getParameters_dispatch.parvec}
\alias{getParameters_dispatch.prdframe}
\alias{getParameters_dispatch.prdlist}
\alias{getParameters_dispatch.eqnlist}
\alias{getParameters_dispatch.eventlist}
\title{Extract the parameters of an object}
\usage{
getParameters(..., conditions = NULL)

getParameters_dispatch(x, ...)

\method{getParameters_dispatch}{odemodel}(x, conditions = NULL, ...)

\method{getParameters_dispatch}{fn}(x, conditions = NULL, ...)

\method{getParameters_dispatch}{parvec}(x, conditions = NULL, ...)

\method{getParameters_dispatch}{prdframe}(x, conditions = NULL, ...)

\method{getParameters_dispatch}{prdlist}(x, conditions = NULL, ...)

\method{getParameters_dispatch}{eqnlist}(x, conditions = NULL, ...)

\method{getParameters_dispatch}{eventlist}(x, conditions = NULL, ...)
}
\arguments{
\item{...}{objects from which the parameters should be extracted}

\item{conditions}{character vector specifying the conditions to
which \code{getParameters} is restricted}

\item{x}{object from which the parameters are extracted}
}
\value{
The parameters in a format that depends on the class of \code{x}.
}
\description{
Extract the parameters of an object
}
