% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classes.R
\name{getParameters}
\alias{getParameters}
\alias{getParameters.fn}
\alias{getParameters.list}
\alias{getParameters.parvec}
\alias{getParameters.prdframe}
\title{Extract the parameters of an object}
\usage{
getParameters(x, ...)

\method{getParameters}{fn}(x, ...)

\method{getParameters}{parvec}(x, ...)

\method{getParameters}{prdframe}(x, ...)

\method{getParameters}{list}(x, ...)
}
\arguments{
\item{x}{object from which the parameters should be extracted}

\item{...}{additional arguments (not used right now)}
}
\value{
The parameters in a format that depends on the class of \code{x}.
}
\description{
Extract the parameters of an object
}

