% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{getParameters_est.grid}
\alias{getParameters_est.grid}
\title{Extract parameter names from est.grid}
\usage{
getParameters_est.grid(est.grid)
}
\arguments{
\item{est.grid}{data.frame(parname, partask, ids...)}
}
\value{
character of outer parameter names
}
\description{
Extract parameter names from est.grid
}
\examples{
eg_good <- stats::setNames(data.frame(parname = "d", partask = "Cond_specific", "1" = NA, "2" = "dummy", "3" = "dummy", stringsAsFactors = FALSE), c("parname", "partask", as.character(1:3)))
getParameters_est.grid(eg_good)
}
