% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod2PEtab_Svenja.R
\name{getParGrids}
\alias{getParGrids}
\title{Get est.grid and fixed.grid}
\usage{
getParGrids(mytrafo, mytrafoL, mycondition.grid, SS_pars = NULL)
}
\arguments{
\item{mytrafo}{base trafo}

\item{mytrafoL}{condition specific branched trafo list}

\item{mycondition.grid}{condition.grid}

\item{SS_pars}{parameters determined by steady state}
}
\value{

}
\description{
Get est.grid and fixed.grid
}
