% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blas.R
\name{\%bmm\%}
\alias{\%bmm\%}
\title{Batched Matrix Multiplication}
\usage{
A \%bmm\% B
}
\arguments{
\item{A}{Numeric matrix or 3D array}

\item{B}{Numeric matrix or 3D array}
}
\value{
Numeric 3D array with dim \code{[M, N, B]}
}
\description{
Efficient batched matrix multiplication with BLAS-backend.
}
\section{Supported Contractions}{

\describe{
\item{\code{[M,K,B] x [K,N] -> [M,N,B]}}{Left batched}
\item{\code{[M,K] x [K,N,B] -> [M,N,B]}}{Right batched}
\item{\code{[M,K,B] x [K,N,B] -> [M,N,B]}}{Both batched}
}
}

\keyword{internal}
