% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blas.R
\name{\%bmm\%}
\alias{\%bmm\%}
\title{Batched Matrix Multiplication}
\usage{
A \%bmm\% B
}
\arguments{
\item{A}{Numeric matrix or 3D array}

\item{B}{Numeric matrix or 3D array}
}
\value{
Numeric 3D array with dim \code{[B, M, N]}
}
\description{
Efficient batched matrix multiplication with BLAS-backend.
Batch index is the FIRST dimension.
}
\section{Supported Contractions}{

\describe{
\item{\code{[B,M,K] x [K,N] -> [B,M,N]}}{Left batched}
\item{\code{[M,K] x [B,K,N] -> [B,M,N]}}{Right batched}
\item{\code{[B,M,K] x [B,K,N] -> [B,M,N]}}{Both batched}
}
}

