% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{gridlist}
\alias{gridlist}
\title{Collect grids in list}
\usage{
gridlist(est.grid, fix.grid)
}
\arguments{
\item{est.grid, fix.grid}{data.tables, will be coerced to one}
}
\value{
list of the grid
}
\description{
Ensure all tables are data.tables
}
