% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLinterface.R
\name{import_sbml}
\alias{import_sbml}
\title{Import an SMBL model}
\usage{
import_sbml(modelpath, amicipath = NULL)
}
\arguments{
\item{modelpath}{Path to the sbml file}

\item{amicipath}{Path to your amici-python-installation, e.g.: AMICIPATH/python}
}
\value{
list of eqnlist, parameters and inits
}
\description{
Requires AMICI https://github.com/ICB-DCM/AMICI/ and dependencies to be installed on your system
Big thanks go to Daniel Weindl!
}
