% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{indiv2Classic_compileTrafo}
\alias{indiv2Classic_compileTrafo}
\title{Compile classic trafo}
\usage{
indiv2Classic_compileTrafo(
  trafoL,
  .compiledFolder = file.path("CompiledObjects")
)
}
\arguments{
\item{trafoL}{output of \link{indiv2Classic_trafo}}

\item{.compiledFolder}{Folder for compiled objects}
}
\value{

}
\description{
Compile classic trafo
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
