% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trustL1.R
\name{l1norm}
\alias{l1norm}
\title{Compute the L1 norm of residuals}
\usage{
l1norm(nout)
}
\arguments{
\item{nout}{data.frame (result of \link{res})}
}
\value{
list with entries value (numeric, the weighted residual sum of squares),
gradient (numeric, gradient) and
hessian (matrix of type numeric).
}
\description{
Compute the L1 norm of residuals
}

