% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{loadDLL}
\alias{loadDLL}
\title{Load shared object for a dMod object}
\usage{
loadDLL(...)
}
\arguments{
\item{...}{objects of class prdfn, obsfn, parfn, objfn, ...}
}
\description{
Usually when restarting the R session, although all objects are saved in
the workspace, the dynamic libraries are not linked any more. \code{loadDLL}
is a wrapper for \code{dyn.load} that uses the "modelname" attribute of
dMod objects like prediction functions, observation functions, etc. to
load the corresponding shared object.
}
