% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{make_pars}
\alias{make_pars}
\title{Extract individualized parameters for one ID}
\usage{
make_pars(parsouter, fixedouter, condition, est.grid, fixed.grid)
}
\arguments{
\item{parsouter, fixedouter}{Arguments which are passed to the function returned by \link{normIndiv}}

\item{condition}{The specific condition for which the parameters should be extracted}

\item{est.grid, fixed.grid}{Lookup tables.}
}
\value{
list(pars, fixed) to put into prd0
}
\description{
From the outer individualized parsouter and fixedouter, get the generalized parsouter0, fixedouter0 to supply to prd0
}
\details{
Catch the following two cases:
\itemize{
\item Entry in est.grid is NA                 => Parameter value present in fixed.grid
\item Parameter name not present in parsouter => Parameter value present in fixed_outer
}
}
