% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModFrameClass.R
\name{mutatedMod.frame}
\alias{mutatedMod.frame}
\title{A version of dplyr::mutate}
\usage{
mutatedMod.frame(dMod.frame, ..., keepCalls = F)
}
\arguments{
\item{dMod.frame}{A dMod.frame}

\item{...}{Expressions going to mutate()}

\item{keepCalls}{Should the dots ... be recorded?}
}
\description{
This is basically dplyr::mutate with two differences
1. It groups the tibble \link{rowwise} before mutating
2. It allows to store the calls. This is intended for use when your objects are
not standard transformations, such as \code{prd = g*x*p} but more complicated and
you want to keep a record of what you did.

If the result of your ... expressions is not atomic, make sure to wrap your
expression in \code{list()}
}
\examples{
\dontrun{
mytbl <- tibble(a = 1:2, b = letters[1:2]) \%>\%
mutatedMod.frame(e = paste0(a,b), keepCalls = T) \%>\%
mutatedMod.frame(f = paste0(e, "=", a, "*", b), keepCalls = T) \%>\%
mutatedMod.frame(e = paste0(a,"*", b), keepCalls = T)

mytbl$.calls
}
}
