% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objClass.R
\name{nll}
\alias{nll}
\title{Compute the negative log-likelihood}
\usage{
nll(
  nout,
  pars,
  deriv,
  opt.BLOQ = "M3",
  opt.hessian = c(ALOQ_part1 = TRUE, ALOQ_part2 = TRUE, ALOQ_part3 = TRUE, BLOQ_part1 =
    TRUE, BLOQ_part2 = TRUE, BLOQ_part3 = TRUE, PD = TRUE)
)
}
\arguments{
\item{nout}{data.frame (result of \link{res}) or object of class \link{res}.}

\item{pars}{Example named vector of outer parameters to construct the objlist}

\item{deriv}{TRUE or FALSE}

\item{opt.BLOQ}{see \link{normIndiv}}

\item{opt.hessian}{see \link{normIndiv}}
}
\value{
list with entries value (numeric, the weighted residual sum of squares),
gradient (numeric, gradient) and
hessian (matrix of type numeric).
}
\description{
Gaussian Log-likelihood. Supports NONMEM-like BLOQ handling methods M1, M3 and M4 and estimation of error models
}
