% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objClass.R
\name{nll_ALOQ}
\alias{nll_ALOQ}
\title{Non-linear log likelihood for the ALOQ part of the data}
\usage{
nll_ALOQ(
  nout,
  derivs,
  derivs.err,
  opt.BLOQ = c("M3", "M4NM", "M4BEAL", "M1"),
  opt.hessian = c(ALOQ_part1 = TRUE, ALOQ_part2 = TRUE, ALOQ_part3 = TRUE)
)
}
\arguments{
\item{nout}{output of \code{\link[=res]{res()}}}

\item{derivs, derivs.err}{attributes of output of \code{\link[=res]{res()}}}

\item{opt.BLOQ}{Character denoting the method to deal with BLOQ data}

\item{opt.hessian}{Named logical vector to include or exclude
various non-convex summands of the hessian matrix}
}
\description{
Non-linear log likelihood for the ALOQ part of the data
}
