% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objClass.R
\name{nll_ALOQ}
\alias{nll_ALOQ}
\title{Non-linear log likelihood for the ALOQ part of the data}
\usage{
nll_ALOQ(
  nout,
  derivs,
  derivs.err,
  deriv2 = NULL,
  deriv2.err = NULL,
  opt.BLOQ = c("M3", "M4NM", "M4BEAL", "M1"),
  opt.hessian = c(ALOQ_part1 = TRUE, ALOQ_part2 = TRUE, ALOQ_part3 = TRUE),
  bessel.correction = 1
)
}
\arguments{
\item{nout}{output of \code{\link[=res]{res()}}}

\item{derivs, derivs.err}{matrix of first derivatives}

\item{deriv2, deriv2.err}{array of second derivatives}

\item{opt.BLOQ}{Character denoting the method to deal with BLOQ data}

\item{opt.hessian}{Named logical vector to include or exclude
various non-convex summands of the hessian matrix}

\item{bessel.correction}{Numeric. Bessel correction factor applied to the
weighted residuals. Default is 1 (no correction).}
}
\description{
Non-linear log likelihood for the ALOQ part of the data
}
