% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objClass.R
\name{normL2}
\alias{normL2}
\title{L2 norm between data and model prediction}
\usage{
normL2(
  data,
  x,
  errmodel = NULL,
  times = NULL,
  attr.name = "data",
  use.bessel = !is.null(errmodel),
  cores = 1L
)
}
\arguments{
\item{data}{Object of class \link{datalist}.}

\item{x}{Object of class \link{prdfn}.}

\item{errmodel}{Optional object of class \link{obsfn}. The error model may be
defined only for a subset of conditions.}

\item{times}{Optional numeric vector of additional time points at which the
prediction function is evaluated. If NULL, time points are taken from the
data. Event times should be included here if the prediction model uses events.}

\item{attr.name}{Character string. The objective value is additionally returned
as an attribute with this name.}

\item{use.bessel}{Logical. If TRUE and an error model is provided, applies a
global Bessel correction to variance estimates to account for finite-sample
bias. Defaults to TRUE if an error model is supplied, FALSE otherwise.}

\item{cores}{Integer. Number of CPU cores used for parallel evaluation over
conditions. Must be >= 1. Parallelization is configured once when the
objective function is created.}
}
\value{
An object of class \code{objfn}, i.e. a function
\code{obj(pars, fixed, deriv, env)} returning an \link{objlist}.
}
\description{
Creates an objective function for parameter estimation based on the
(negative log-likelihood) L2 norm between observed data and model predictions.
The returned objective function can be used with optimizers such as
\link{mstrust} and supports aggregation over multiple experimental conditions.
}
\details{
Objective functions can be combined using the \code{+} operator, see
\link{sumobjfn}.

The Bessel correction is applied globally across all conditions and is given by
\deqn{\sqrt{n / (n - p)}}
where \eqn{n} is the total number of data points and \eqn{p} is the number of
structural (non-error-model) parameters.

Parallelization is performed over experimental conditions if
\code{cores > 1}. The number of cores is fixed when calling \code{normL2()}
and cannot be changed at evaluation time.
}
\examples{
\dontrun{
  ## Generate a prediction function
  
  times <- 0:5
  grid <- data.frame(name = "A", time = times, row.names = paste0("p", times))
  x <- Xd(grid, condition = "C1")
  
  pars <- structure(rep(0, nrow(grid)), names = row.names(grid))
  
  ## Simulate data
  data.list <- lapply(1:3, function(i) {
    prediction <- x(times, pars + rnorm(length(pars), 0, 1))
    cbind(wide2long(prediction), sigma = 1)
  })
  
  data <- as.datalist(do.call(rbind, data.list))
  
  ## Generate objective function based on data and model
  ## Then fit the data and plot the result
  obj <- normL2(data, x)
  myfit <- trust(obj, pars, rinit = 1, rmax = 10)
  plot(x(times, myfit$argument), data)
}
}
