% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objClass.R
\name{normL2}
\alias{normL2}
\title{L2 norm between data and model prediction}
\usage{
normL2(
  data,
  x,
  errmodel = NULL,
  times = NULL,
  attr.name = "data",
  use.bessel = ifelse(!is.null(errmodel), TRUE, FALSE)
)
}
\arguments{
\item{data}{Object of class \link{datalist}.}

\item{x}{Object of class \link{prdfn}.}

\item{errmodel}{Optional object of class \link{obsfn}. The error model does not
need to be defined for all conditions.}

\item{times}{Numeric vector of additional time points where the prediction
function is evaluated. If NULL, time points are extracted from the data.
If the prediction function uses events, event times should be provided here.}

\item{attr.name}{Character string. The objective value is additionally returned
as an attribute with this name.}

\item{use.bessel}{Logical. If TRUE and an error model is present, applies a
global Bessel correction to variance estimates to account for finite-sample
bias. Default is TRUE if an error model is provided, FALSE otherwise.}
}
\value{
An object of class \code{objfn}, i.e. a function
\code{obj(pars, fixed, deriv, deriv2, conditions, env, cores)} returning an
\link{objlist}.
}
\description{
For parameter estimation and optimization, an objective function is needed.
\code{normL2} returns an objective function based on the (negative log-likelihood)
L2 norm between data and model prediction. The resulting objective function
can be used for optimization with the trust optimizer, see \link{mstrust}.
}
\details{
Objective functions can be combined using the \code{+} operator, see \link{sumobjfn}.

The Bessel correction is applied globally across all conditions and is given by
\deqn{\sqrt{n / (n - p)}}
where \eqn{n} is the total number of data points and \eqn{p} the number of
structural (non-error-model) parameters.

Parallel evaluation over conditions is supported via the \code{cores} argument
of the returned objective function. Parallelization is only applied if
\code{cores > 1}. A value of \code{cores <= 0} is not allowed.
}
\examples{
\dontrun{
  ## Generate a prediction function
  
  times <- 0:5
  grid <- data.frame(name = "A", time = times, row.names = paste0("p", times))
  x <- Xd(grid, condition = "C1")
  
  pars <- structure(rep(0, nrow(grid)), names = row.names(grid))
  
  ## Simulate data
  data.list <- lapply(1:3, function(i) {
    prediction <- x(times, pars + rnorm(length(pars), 0, 1))
    cbind(wide2long(prediction), sigma = 1)
  })
  
  data <- as.datalist(do.call(rbind, data.list))
  
  ## Generate objective function based on data and model
  ## Then fit the data and plot the result
  obj <- normL2(data, x)
  myfit <- trust(obj, pars, rinit = 1, rmax = 10)
  plot(x(times, myfit$argument), data)
}
}
