% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objClass.R
\name{normL2}
\alias{normL2}
\title{L2 norm between data and model prediction}
\usage{
normL2(
  data,
  x,
  errmodel = NULL,
  times = NULL,
  attr.name = "data",
  use.bessel = ifelse(!is.null(errmodel), TRUE, FALSE)
)
}
\arguments{
\item{data}{object of class \link{datalist}}

\item{x}{object of class \link{prdfn}}

\item{errmodel}{object of class \link{obsfn}. \code{errmodel} does not need to be defined for all conditions.}

\item{times}{numeric vector, additional time points where the prediction function is 
evaluated. If NULL, time points are extracted from the datalist solely. If the prediction
function makes use of events, hand over event \code{times} here.}

\item{attr.name}{character. The constraint value is additionally returned in an 
attribute with this name}

\item{use.bessel}{logical. If TRUE and an error model is present, applies Bessel correction
to account for the bias in maximum likelihood estimation of variance parameters. 
The correction factor is sqrt(n/(n-p)) where n is the total number of data points and 
p is the number of structural (non-error-model) parameters. Default is TRUE if errmodel 
is provided, FALSE otherwise.}
}
\value{
Object of class \code{objfn}, i.e. a function 
\code{obj(..., fixed, deriv, conditions, env)} that returns an objective list,
\link{objlist}.
}
\description{
For parameter estimation and optimization, an objective function
is needed. \code{normL2} returns an objective function for the L2 norm of
data and model prediction. The resulting objective function can be used for
optimization with the trust optimizer, see \link{mstrust}.
}
\details{
Objective functions can be combined by the "+" operator, see \link{sumobjfn}.

The Bessel correction addresses the downward bias in maximum likelihood estimates of 
variance parameters. When use.bessel = TRUE, the correction is applied globally across 
all conditions based on the total number of data points and structural parameters, 
consistent with NONMEM's approach where only THETA parameters (not SIGMA) contribute 
to degrees of freedom.
}
\examples{
## Generate a prediction function

times <- 0:5
grid <- data.frame(name = "A", time = times, row.names = paste0("p", times))
x <- Xd(grid, condition = "C1")

pars <- structure(rep(0, nrow(grid)), names = row.names(grid))

## Simulate data
data.list <- lapply(1:3, function(i) {
  prediction <- x(times, pars + rnorm(length(pars), 0, 1))
  cbind(wide2long(prediction), sigma = 1)
})

data <- as.datalist(do.call(rbind, data.list))

## Generate objective function based on data and model
## Then fit the data and plot the result
obj <- normL2(data, x)
myfit <- trust(obj, pars, rinit = 1, rmax = 10)
plot(x(times, myfit$argument), data)
}
