% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{objframe}
\alias{objframe}
\title{Objective frame}
\usage{
objframe(mydata, deriv = NULL, deriv.err = NULL)
}
\arguments{
\item{mydata}{data.frame as being generated by \link{res}.}

\item{deriv}{matrix of the derivatives of the residuals with respect to parameters.}

\item{deriv.err}{matrix of the derivatives of the error model.}
}
\value{
An object of class \code{objframe}, i.e. a data frame with attribute "deriv".
}
\description{
An objective frame is supposed to store the residuals of a model prediction
with respect to a data frame.
}
