% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{objlist}
\alias{objlist}
\title{Generate objective list}
\usage{
objlist(value, gradient, hessian)
}
\arguments{
\item{value}{numeric of length 1}

\item{gradient}{named numeric}

\item{hessian}{matrix with rownames and colnames according to gradient names}
}
\value{
Object of class \code{objlist}
}
\description{
An objective list contains an objective value, a gradient, and a Hessian matrix.

Objective lists can contain additional numeric attributes that are preserved or
combined with the corresponding attributes of another objective list when
both are added by the "+" operator, see \link{sumobjlist}.

Objective lists are returned by objective functions as being generated
by \link{normL2}, \link{constraintL2}, \link{priorL2} and \link{datapointL2}.
}
