% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDenial.R
\name{optionsLSODES}
\alias{optionsLSODES}
\alias{optionsLSODE}
\alias{optionsLSODA}
\title{Get default arguments for integrators}
\usage{
optionsLSODES(
  rtol = 1e-06,
  atol = 1e-06,
  jacvec = NULL,
  sparsetype = "sparseint",
  nnz = NULL,
  inz = NULL,
  rootfunc = NULL,
  verbose = FALSE,
  nroot = 0,
  tcrit = NULL,
  hmin = 0,
  hmax = NULL,
  hini = 0,
  ynames = TRUE,
  maxord = NULL,
  maxsteps = 5000,
  lrw = NULL,
  liw = NULL
)

optionsLSODE(
  rtol = 1e-06,
  atol = 1e-06,
  jacfunc = NULL,
  jactype = "fullint",
  mf = NULL,
  rootfunc = NULL,
  verbose = FALSE,
  nroot = 0,
  tcrit = NULL,
  hmin = 0,
  hmax = NULL,
  hini = 0,
  ynames = TRUE,
  maxord = NULL,
  bandup = NULL,
  banddown = NULL,
  maxsteps = 5000,
  rpar = NULL,
  ipar = NULL,
  nout = 0,
  outnames = NULL
)

optionsLSODA(
  rtol = 1e-06,
  atol = 1e-06,
  jacfunc = NULL,
  jactype = "fullint",
  rootfunc = NULL,
  verbose = FALSE,
  nroot = 0,
  tcrit = NULL,
  hmin = 0,
  hmax = NULL,
  hini = 0,
  ynames = TRUE,
  maxordn = 12,
  maxords = 5,
  bandup = NULL,
  banddown = NULL,
  maxsteps = 5000
)
}
\value{
List of arguments
}
\description{
Get default arguments for integrators
}
