% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parameters.R
\name{P}
\alias{P}
\title{Parameter transformation}
\usage{
P(P, parameters = NULL)
}
\arguments{
\item{P}{named character vector. Names correspond to the parameter names to be fed into
the model. The values of P are equations that express the parameters in terms of other
parameters}

\item{parameters}{character vector. Optional. If given, the generated parameter
transformation returns values for each element in \code{parameters}. If elements of
\code{parameters} are not in \code{names(P)} the identity transformation is assumed.}

\item{P}{named character vector. Names correspond to the parameter names to be fed into
the model. The values of P are equations that express the parameters in terms of other
parameters}

\item{parameters}{character vector. Optional. If given, the generated parameter
transformation returns values for each element in \code{parameters}. If elements of
\code{parameters} are not in \code{names(P)} the identity transformation is assumed.}
}
\value{
a function \code{p(P)} representing the parameter transformation. The result of p(P)
contains an attribute "deriv" which is the jacobian of the parameter transformation
evaluated at \code{P}.
Parameter transformation

a function \code{p(P)} representing the parameter transformation. The result of p(P)
contains an attribute "deriv" which is the jacobian of the parameter transformation
evaluated at \code{P}.
}
\description{
Parameter transformation
}

