% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{pdIndiv_updateControls}
\alias{pdIndiv_updateControls}
\title{Quickly set "controls" amd update all high-level functions}
\usage{
pdIndiv_updateControls(
  pd,
  optionsOde = NULL,
  optionsSens = NULL,
  objtimes = NULL
)
}
\arguments{
\item{pd}{}

\item{objtimes}{}

\item{rtol}{}

\item{atol}{}

\item{maxsteps}{}
}
\value{

}
\description{
Quickly set "controls" amd update all high-level functions
}
\examples{
optionsOde <- list(method = "lsoda", rtol = rtol, atol = atol, maxsteps = maxsteps)
optionsSens <- list(method = "lsodes", rtol = rtol, atol = atol, maxsteps = maxsteps)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
