% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCombined}
\alias{plotCombined}
\title{Plot a list of model predictions and a list of data points in a combined plot}
\usage{
plotCombined(prediction, data, ...)
}
\arguments{
\item{prediction}{Named list of matrices or data.frames, usually the output of a prediction function
as generated by \link{Xs}.}

\item{data}{Named list of data.frames as being used in \link{res}, i.e. with columns \code{name}, \code{time},
\code{value} and \code{sigma}.}

\item{...}{Further arguments going to \code{subset}.}
}
\value{
A plot object of class \code{ggplot}.
}
\description{
Plot a list of model predictions and a list of data points in a combined plot
}
\details{
The data.frame being plotted has columns \code{time}, \code{value}, \code{sigma},
\code{name} and \code{condition}.
}

