% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotData}
\alias{plotData}
\title{Plot a list data points}
\usage{
plotData(data, ...)
}
\arguments{
\item{data}{Named list of data.frames as being used in \link{res}, i.e. with columns \code{name}, \code{time},
\code{value} and \code{sigma}.}

\item{...}{Further arguments going to \code{subset}.}
}
\value{
A plot object of class \code{ggplot}.
}
\description{
Plot a list data points
}
\details{
The data.frame being plotted has columns \code{time}, \code{value}, \code{sigma},
\code{name} and \code{condition}.
}

