% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPars}
\alias{plotPars}
\title{Plot parameter values for a fitlist}
\usage{
plotPars(myparframe, whichFits = 1:nrow(myparframe))
}
\arguments{
\item{myparframe}{parameter frame as obtained by as.parframe(mstrust)}

\item{whichFits}{indexlist e.g. 1:10}
}
\description{
Plot parameter values for a fitlist
}

