% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPars}
\alias{plotPars}
\title{Plot parameter values for a fitlist}
\usage{
plotPars(x, ...)
}
\arguments{
\item{x}{parameter frame as obtained by as.parframe(mstrust)}

\item{...}{arguments for subsetting of x}
}
\description{
Plot parameter values for a fitlist
}

