% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parClass.R, R/plots.R
\name{plotPars.parframe}
\alias{plotPars.parframe}
\alias{plotPars}
\title{Plot parameter values for a fitlist}
\usage{
\method{plotPars}{parframe}(x, tol = 1, ...)

plotPars(x, ...)
}
\arguments{
\item{x}{parameter frame as obtained by as.parframe(mstrust)}

\item{tol}{maximal allowed difference between neighboring objective values
to be recognized as one.}

\item{...}{arguments for subsetting of x}
}
\description{
Plot parameter values for a fitlist
}
