% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModFrameClass.R, R/plots.R
\name{plotPaths.tbl_df}
\alias{plotPaths.tbl_df}
\alias{plotPaths}
\title{Profile likelihood: plot of the parameter paths.}
\usage{
plotPaths.tbl_df(dMod.frame, hypothesis = 1, ...)

plotPaths(
  profs,
  ...,
  whichPar = NULL,
  sort = FALSE,
  relative = TRUE,
  scales = "fixed"
)
}
\arguments{
\item{hypothesis}{numeric, can be longer than 1}

\item{...}{arguments going to subset}

\item{profs}{profile or list of profiles as being returned by \link{profile}}

\item{whichPar}{Character or index vector, indicating the parameters that are taken as possible reference (x-axis)}

\item{sort}{Logical. If paths from different parameter profiles are plotted together, possible
combinations are either sorted or all combinations are taken as they are.}

\item{relative}{logical indicating whether the origin should be shifted.}

\item{scales}{character, either \code{"free"} or \code{"fixed"}.}
}
\value{
A plot object of class \code{ggplot}.
}
\description{
Profile likelihood: plot of the parameter paths.
}
\details{
See \link{profile} for examples.
}
