% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPaths}
\alias{plotPaths}
\title{Profile likelihood: plot of the parameter paths.}
\usage{
plotPaths(profs, ..., whichPar = NULL, sort = FALSE, relative = TRUE,
  scales = "fixed")
}
\arguments{
\item{profs}{profile or list of profiles as being returned by \link{profile}}

\item{...}{arguments going to subset}

\item{whichPar}{Character or index vector, indicating the parameters that are taken as possible reference (x-axis)}

\item{sort}{Logical. If paths from different parameter profiles are plotted together, possible
combinations are either sorted or all combinations are taken as they are.}

\item{relative}{logical indicating whether the origin should be shifted.}

\item{scales}{character, either \code{"free"} or \code{"fixed"}.}
}
\value{
A plot object of class \code{ggplot}.
}
\description{
Profile likelihood: plot of the parameter paths.
}
\details{
See \link{profile} for examples.
}

