% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPrediction}
\alias{plotPrediction}
\title{Plot a list of model predictions}
\usage{
plotPrediction(prediction, ...)
}
\arguments{
\item{prediction}{Named list of matrices or data.frames, usually the output of a prediction function
as generated by \link{Xs}.}

\item{...}{Further arguments going to \code{subset}.}
}
\value{
A plot object of class \code{ggplot}.
}
\description{
Plot a list of model predictions
}
\details{
The data.frame being plotted has columns \code{time}, \code{value}, \code{name} and \code{condition}.
}

