% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotProfile}
\alias{plotProfile}
\title{Profile likelihood plot}
\usage{
plotProfile(..., maxvalue = 5)
}
\arguments{
\item{...}{Lists of profiles as being returned by \link{profile}.}

\item{maxvalue}{Numeric, the value where profiles are cut off.}
}
\value{
A plot object of class \code{ggplot}.
}
\description{
Profile likelihood plot
}

