% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModFrameClass.R, R/parClass.R, R/plots.R
\name{plotProfile.tbl_df}
\alias{plotProfile.tbl_df}
\alias{plotProfile.parframe}
\alias{plotProfile.list}
\alias{plotProfile}
\title{Profile likelihood plot}
\usage{
\method{plotProfile}{tbl_df}(dMod.frame, hypothesis = 1, ...)

\method{plotProfile}{parframe}(profs, ..., maxvalue = 5, parlist = NULL)

\method{plotProfile}{list}(profs, ..., maxvalue = 5, parlist = NULL)

plotProfile(profs, ...)
}
\arguments{
\item{hypothesis}{numeric, can be longer than 1}

\item{...}{logical going to subset before plotting.}

\item{profs}{Lists of profiles as being returned by \link{profile}.}

\item{maxvalue}{Numeric, the value where profiles are cut off.}

\item{parlist}{Matrix or data.frame with columns for the parameters to be added to the plot as points.
If a "value" column is contained, deltas are calculated with respect to lowest chisquare of profiles.}
}
\value{
A plot object of class \code{ggplot}.
}
\description{
Profile likelihood plot
}
\details{
See \link{profile} for examples.
}
