% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotProfile}
\alias{plotProfile}
\title{Profile likelihood plot}
\usage{
plotProfile(profs, ..., maxvalue = 5, parlist = NULL)
}
\arguments{
\item{profs}{Lists of profiles as being returned by \link{profile}.}

\item{...}{logical going to subset before plotting.}

\item{maxvalue}{Numeric, the value where profiles are cut off.}

\item{parlist}{Matrix or data.frame with columns for the parameters to be added to the plot as points.
If a "value" column is contained, deltas are calculated with respect to lowest chisquare of profiles.}
}
\value{
A plot object of class \code{ggplot}.
}
\description{
Profile likelihood plot
}
\details{
See \link{profile} for examples.
}

