% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsSvenja.R
\name{plotProfilesAndPaths}
\alias{plotProfilesAndPaths}
\title{Profile likelihood: plot profiles along with their parameter paths}
\usage{
plotProfilesAndPaths(profs, whichpars, npars = 5, normalizePaths = FALSE)
}
\arguments{
\item{profs}{Lists of profiles as being returned by \link{profile}.}

\item{whichpars}{Character vector of parameter names for which the profile paths should be generated.}

\item{npars}{Numeric vector of colored and named parameter paths.}

\item{normalizePaths}{Logical indicating whether the paths should be normalized to absolute values of 1. Default \code{FALSE}; \code{TRUE} only useful in corner cases when you know why to do so.}
}
\value{
A plot object of class \code{ggplot}.
}
\description{
Profile likelihood: plot profiles along with their parameter paths
}
\examples{
\dontrun{
 plotProfilesAndPaths(myprofiles, c("mypar1", "mypar2"), npars = 5) 
}
}
\author{
Svenja Kemmer, \email{svenja.kemmer@fdm.uni-freiburg.de}
}
