% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotValues}
\alias{plotValues}
\title{Plotting objective values of a collection of fits}
\usage{
plotValues(x, tol = 1, ...)
}
\arguments{
\item{x}{data.frame with columns "value", "converged" and "iterations", e.g. 
a \link{parframe}.}

\item{tol}{maximal allowed difference between neighboring objective values
to be recognized as one.}

\item{...}{arguments for subsetting of x}
}
\description{
Plotting objective values of a collection of fits
}
