% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModFrameClass.R, R/parClass.R, R/plots.R
\name{plotValues.tbl_df}
\alias{plotValues.tbl_df}
\alias{plotValues.parframe}
\alias{plotValues}
\title{Plotting objective values of a collection of fits}
\usage{
\method{plotValues}{tbl_df}(dMod.frame, hypothesis = 1, ..., tol = 1)

\method{plotValues}{parframe}(x, tol = 1, ...)

plotValues(x, ...)
}
\arguments{
\item{...}{arguments for subsetting of x}

\item{tol}{maximal allowed difference between neighboring objective values
to be recognized as one.}

\item{x}{data.frame with columns "value", "converged" and "iterations", e.g. 
a \link{parframe}.}
}
\description{
Plotting objective values of a collection of fits
}
