% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objClass.R
\name{+.objlist}
\alias{+.objlist}
\alias{sumobjlist}
\title{Add two lists element by element}
\usage{
\method{+}{objlist}(out1, out2)
}
\arguments{
\item{out1}{List of numerics or matrices}

\item{out2}{List with the same structure as out1 (there will be no warning when mismatching)}
}
\value{
List of length of out1.
}
\description{
Add two lists element by element
}
\details{
If out1 has names, out2 is assumed to share these names. Each element of the list out1
is inspected. If it has a \code{names} attributed, it is used to do a matching between out1 and out2.
The same holds for the attributed \code{dimnames}. In all other cases, the "+" operator is applied
the corresponding elements of out1 and out2 as they are.
}

