% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDenial.R
\name{predtimes}
\alias{predtimes}
\title{Get a vector of linearly spaced time points for prediction}
\usage{
predtimes(datatimes, eventtimes = NULL, Nobjtimes = 100)
}
\arguments{
\item{datatimes}{}

\item{eventtimes}{}

\item{Nobjtimes}{number of time points in total}
}
\value{
vector of time points
}
\description{
for nice plots
}
\examples{
predtimes(c(30,60,60,90))
predtimes(c(-30,60,60,90))
}
