% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqnClass.R
\name{print.eqnlist}
\alias{print.eqnlist}
\title{Print or pander equation list}
\usage{
\method{print}{eqnlist}(x, pander = FALSE, ...)
}
\arguments{
\item{x}{object of class \link{eqnlist}}

\item{pander}{logical, use pander for output (used with R markdown)}

\item{...}{additional arguments}
}
\description{
Print or pander equation list
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}

Daniel Kaschek, \email{daniel.kaschek@physik.uni-freiburg.de}
}
