% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parClass.R
\name{print.parvec}
\alias{print.parvec}
\title{Pretty printing for a parameter vector}
\usage{
\method{print}{parvec}(x, ...)
}
\arguments{
\item{x}{object of class \code{parvec}}

\item{...}{not used yet.}
}
\description{
Pretty printing for a parameter vector
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}

