% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toolsMirjam.R
\name{print0}
\alias{print0}
\title{Print without attributes}
\usage{
print0(x, list_attributes = TRUE)
}
\arguments{
\item{x}{The object to be printed out}

\item{list_attributes}{prints a list of attribute names if TRUE (default=TRUE)}
}
\description{
Print without attributes
}
\details{
To suppress the printout of attributes like "deriv".
}

