% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{reduceReplicates}
\alias{reduceReplicates}
\title{Reduce replicated measurements to mean and standard deviation}
\format{
The following columns are mandatory for the data frame:
\describe{
 \item{name}{Name of the observed species.}
 \item{time}{Measurement time point.}
 \item{value}{Measurement value.}
 \item{condition}{The condition under which the observation was made.}
}

In addition to these columns, any number of columns can follow to allow a
fine-grained definition of conditions. The values of all columns named in
\code{select} are then merged to get the set of conditions.
}
\usage{
reduceReplicates(data, select = "condition", datatrans = NULL)
}
\arguments{
\item{data}{A data frame containing the measurements. See Format for details.}

\item{select}{Names of the columns in the data frame used to define
conditions, see Details.}

\item{datatrans}{Character vector describing a function to transform data.
Use \kbd{x} to refer to data.}
}
\value{
A data frame of the following variables:
\describe{
 \item{time}{Measurement time point.}
 \item{name}{Name of the observed species.}
 \item{value}{Mean of replicates.}
 \item{sigma}{Standard error of the mean, NA for single measurements.}
 \item{n}{The number of replicates reduced.}
 \item{condition}{The condition for which the value and sigma were calculated. If
       more than one column was used to define the condition, this variable
       holds the effective condition which is the combination of all applied
       single conditions. }
}
}
\description{
Obtain the mean and standard deviation from replicates per condition.
}
\details{
Experiments are usually repeated multiple times possibly under different
conditions leading to replicated measurements. The column "condition" in the
data allows grouping the data by their condition. However, sometimes, a more
fine-grained grouping is desirable. In this case, any number of additional
columns can be appended to the data. These columns are referred to as
"condition identifiers". Which of the condition identifiers are used for
grouping is user-defined by specifying their names in \code{select}. The mandatory
column "condition" is always used. The total set of different conditions is
thus defined by all combinations of values occurring in the selected condition
identifiers. The replicates of each condition are then reduced to mean and 
standard deviation. New condition names are derived by merging all conditions 
which were used in mean and standard deviation.
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}

Simon Beyer, \email{simon.beyer@fdm.uni-freiburg.de}
}
