% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{res}
\alias{res}
\title{Compare data and model prediction by computing residuals}
\usage{
res(data, out)
}
\arguments{
\item{data}{data.frame with name (factor), time (numeric), value (numeric) and sigma (numeric)}

\item{out}{output of ode(), optionally augmented with attributes
"deriv" (output of ode() for the sensitivity equations) and
"parameters" (character vector of parameter names, a subsest of those
contained in the sensitivity equations). If "deriv" is given, also "parameters"
needs to be given.}
}
\value{
data.frame with the original data augmented by columns "prediction" (
numeric, the model prediction), "residual" (numeric, difference between
prediction and data value), "weighted.residual" (numeric, residual devided
by sigma). If "deriv" was given, the returned data.frame has an
attribute "deriv" (data.frame with the derivatives of the residuals with
respect to the parameters).
}
\description{
Compare data and model prediction by computing residuals
}

