% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{res}
\alias{res}
\title{Compute residuals}
\usage{
res(data, out, err = NULL)
}
\arguments{
\item{data}{Data frame with columns time, name, value, sigma, lloq}

\item{out}{Numeric matrix with predictions (first column = time)}

\item{err}{Optional error model predictions}
}
\value{
Object of class objframe with data and derivatives
}
\description{
Compute residuals
}
