% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{res}
\alias{res}
\title{Compare data and model prediction by computing residuals}
\usage{
res(data, out, err = NULL)
}
\arguments{
\item{data}{data.frame with columns: time, name, value, sigma, lloq}

\item{out}{matrix with predictions (first column = time)}

\item{err}{Nullable matrix with error model predictions}
}
\description{
When sigma is given in the data, it always has priority over the sigma in the error model
}
