% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{res_cpp}
\alias{res_cpp}
\title{Compute residuals (Rcpp implementation, optimized)}
\usage{
res_cpp(data, out, err = NULL)
}
\arguments{
\item{data}{DataFrame with columns: time, name, value, sigma, lloq}

\item{out}{NumericMatrix with predictions (first column = time)}

\item{err}{Nullable NumericMatrix with error model predictions}
}
\value{
List with:
\itemize{
\item data: DataFrame with residual columns
\item deriv:  NumericMatrix n_data × n_pars (first derivatives of out)
\item deriv2: NumericArray  n_data × n_pars × n_pars (second derivatives of out)
\item deriv.err:  NumericMatrix n_data × n_pars (first derivatives of err sigma-part)
\item deriv2.err: NumericArray  n_data × n_pars × n_pars (second derivatives of err sigma-part)
}
}
\description{
Compute residuals (Rcpp implementation, optimized)
}
