% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resObjClass.R
\name{resL2}
\alias{resL2}
\title{Residual-based Objective Function for lsqcpp optimization}
\usage{
resL2(data, x, errmodel = NULL, times = NULL, optBLOQ = "none")
}
\arguments{
\item{data}{object of class \link{datalist}}

\item{x}{object of class \link{prdfn}}

\item{errmodel}{object of class \link{obsfn}}

\item{times}{numeric vector, additional time points}

\item{optBLOQ}{Character: "none", "M1", "M3", or "M4NM" for BLOQ handling}
}
\value{
Object of class `resObjfn` - a function that returns a list with:
  \item{residuals}{Numeric vector of weighted residuals plus artificial residuals for log(sigma)}
  \item{jacobian}{Matrix with derivatives of residuals w.r.t. all parameters (model + error)}
}
\description{
Creates an objective function that returns residuals and Jacobians compatible
with lsqcpp C++ optimizers (Gauss-Newton, Levenberg-Marquardt, etc.).

Unlike `normL2()` which returns an objlist with value/gradient/hessian,
`resL2()` returns residuals suitable for least-squares optimization.
}
\details{
The returned function can be used with lsqcpp optimizers or combined with
constraints using `+` operator:

```
obj <- resL2(data, model) + constraintResL2(mu, sigma)
```

**Important:** This function includes artificial residuals for the log(sigma) terms
from the likelihood. For each ALOQ data point, an additional residual 
`sqrt(log(2*pi*sigma^2))` is added to ensure the likelihood is correctly 
represented as a sum of squared residuals.
}
\seealso{
Other dMod interface: 
\code{\link{constraintResL2}()},
\code{\link{datapointResL2}()},
\code{\link{res}()},
\code{\link{resCpp}()}
}
\concept{dMod interface}
