% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{resObjlist}
\alias{resObjlist}
\title{Generate residual object}
\usage{
resObjlist(residuals, jacobian = NULL)
}
\arguments{
\item{residuals}{Numeric vector of residuals}

\item{jacobian}{Matrix with derivatives of residuals w.r.t. parameters (optional)}
}
\value{
Object of class \code{resObjlist}
}
\description{
A residual object contains residuals and an optional Jacobian matrix.
This is the output format for lsqcpp-compatible objective functions.

Residual objects are returned by residual objective functions as being generated
by \link{resL2}, \link{constraintResL2}, and \link{datapointResL2}.
}
