% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolics.R
\name{rref}
\alias{rref}
\title{Transform matrix A into reduced row echelon form 
this function is written along the lines of the rref-matlab function.}
\usage{
rref(A, tol = sqrt(.Machine$double.eps), verbose = FALSE, fractions = FALSE)
}
\arguments{
\item{A}{matrix for which the reduced row echelon form is searched}

\item{tol}{tolerance to find pivots}

\item{verbose}{logical, print verbose information}

\item{fractions}{logical, not used right now.}
}
\value{
a list of two entries is returned; ret[[1]] is the reduced row echelon form of A, ret[[2]] is the index of columns in which a pivot was found
}
\description{
Transform matrix A into reduced row echelon form 
this function is written along the lines of the rref-matlab function.
}
\author{
Malenka Mader, \email{Malenka.Mader@fdm.uni-freiburg.de}
}
