% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRtools.R
\name{run_IQRsysProject}
\alias{run_IQRsysProject}
\title{Runs an IQRsysProject}
\usage{
run_IQRsysProject(proj, ncores = 1, opt.nfits = 10, opt.sd = 1,
  opt.iterlim = 100, FLAGprofileLL = FALSE)
}
\description{
Runs an IQRsysProject
}
