% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scale_fill_dMod}
\alias{scale_fill_dMod}
\title{Standard dMod color scheme}
\usage{
scale_fill_dMod(...)
}
\arguments{
\item{...}{arguments goint to code{scale_color_manual()}}
}
\description{
Standard dMod color scheme
}
