% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toolsMarcus.R
\name{steadyStates}
\alias{steadyStates}
\title{Calculate analytical steady states}
\usage{
steadyStates(model, file, forcings = "", neglect = "", outputFormat = "R",
  sparsifyLevel = 2)
}
\arguments{
\item{model}{Name of the file which the stoechiometric matric in csv format.}

\item{file}{Name of the file to which the steady-state equations are saved.
Read this file with \code{\link{readRDS}}.}

\item{forcings}{Character vector with the names of the forcings}

\item{neglect}{Character vector with the names of species which must not be
used for solving the steady state equations.}

\item{outputFormat}{Define the output format. By default "R" generating dMod
compatible output. To obtain an output appropriate for d2d [1] "M" must be
selected.}

\item{sparsifyLevel}{numeric, set the level to which the stoichiometric matrix is
  simplified by searching for clever linear combinations.}
}
\value{
Character vector of steady-state equations.
}
\description{
Calculate analytical steady states
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}
\references{
[1]
\url{https://bitbucket.org/d2d-development/d2d-software/wiki/Home}
}

