% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toolsBeni.R
\name{symmetryDetection}
\alias{symmetryDetection}
\title{Search for symmetries in the loaded model}
\usage{
symmetryDetection(f, obsvect = NULL, prediction = NULL, initial = NULL,
  ansatz = "uni", pMax = 2, inputs = c(), fixed = c(), cores = 1,
  allTrafos = FALSE)
}
\arguments{
\item{f}{eqnList object or vector containing ODEs}

\item{obsvect}{vector of observation functions}

\item{prediction}{vector containing prediction to be tested}

\item{initial}{vector containing initial values}

\item{ansatz}{type of infinitesimal ansatz used for the analysis (uni, par, multi)}

\item{pMax}{maximal degree of infinitesimal ansatz}

\item{inputs}{specify the input variables}

\item{fixed}{variables to concider fixed}

\item{cores}{maximal number of cores used for the analysis}

\item{allTrafos}{do not remove transformations with a common parameter factor}
}
\description{
Search for symmetries in the loaded model
}

