% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab2dMod.R
\name{updateParscalesToBaseTrafo}
\alias{updateParscalesToBaseTrafo}
\title{Rename parscales to the names needed in the base trafo}
\usage{
updateParscalesToBaseTrafo(parscales, est.grid)
}
\arguments{
\item{parscales}{setNames(PETABPars$parameterScale, PETABPars$parameterId)}

\item{est.grid}{data.table}
}
\value{

}
\description{
Rename parscales to the names needed in the base trafo
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
