% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqnClass.R
\name{x2logx}
\alias{x2logx}
\title{Log-transform variables in an equation vector using SymPy}
\usage{
x2logx(eqnvec, whichVar)
}
\arguments{
\item{eqnvec}{An object of class \code{eqnvec}, representing a set of equations.}

\item{whichVar}{A character vector specifying the variables to be log-transformed.}
}
\value{
An updated \code{eqnvec} object with transformed equations.
}
\description{
Log-transform variables in an equation vector using SymPy
}
\details{
The function applies a logarithmic transformation to the specified variables in the equation vector.
For a variable \code{var}, the transformation is \code{log(var)} and the derivative \code{d/dt log(var)} is replaced by 
\code{(1/exp(log(var))) * d/dt var}, substituting \code{var} with \code{exp(log(var))} in the original equation.
The original variable equations are replaced or removed, and the transformed equations are added with updated names prefixed by \code{log_}.
The equations are simplified using SymPy to ensure mathematical correctness and compactness.
}
\examples{
eqnvec <- as.eqnvec(
  c("-k1*A + k2*B", "k1*A - k2*B"),
  names = c("A", "B")
)
log_transformed_eqns <- x2logx(eqnvec, c("A", "B"))

}
