% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prediction.R
\name{Xf}
\alias{Xf}
\title{Model evaluation without sensitivities.}
\usage{
Xf(func, forcings = NULL, events = NULL, optionsOde = list(method =
  "lsoda"))
}
\arguments{
\item{func}{return value from \code{funC(f)} where \code{f} defines the ODE.}

\item{forcings}{data.frame with columns name (factor), time (numeric) and value (numeric).
The ODE forcings.}

\item{events}{data.frame of events with columns "var" (character, the name of the state to be
affected), "time" (numeric, time point), "value" (numeric, value), "method" (character, either
"replace", "add" or "multiply"). See \link[deSolve]{events}.}

\item{optionsOde}{list with arguments to be passed to odeC() for the ODE integration.}
}
\description{
Interface to get an ODE
into a model function \code{x(times, pars, forcings, events)} returning ODE output.
It is a reduced version of \link{Xs}, missing the sensitivities.
}
\details{
Can be used to integrate additional quantities, e.g. fluxes, by adding them to \code{f}. All quantities that are not initialised by pars
in \code{x(times, pars, forcings, events)} are initialized at 0.
}

