% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prediction.R
\name{Y}
\alias{Y}
\title{Observation functions.}
\usage{
Y(g, f = NULL, states = NULL, parameters = NULL, condition = NULL,
  attach.input = TRUE, compile = FALSE, modelname = NULL,
  verbose = FALSE)
}
\arguments{
\item{g}{Named character vector defining the observation function}

\item{f}{Named character, the underlying ODE}

\item{states}{character vector, alternative definition of "states", usually the names of \code{f}}

\item{parameters}{character vector, alternative definition of the "parameters",
usually the symbols contained in "g" and "f" except for \code{states} and the code word \code{time}.}

\item{condition}{either NULL (generic prediction for any condition) or a character, denoting
the condition for which the function makes a prediction.}

\item{attach.input}{logical, indiating whether the original input should be
returned with the output.}

\item{compile}{Logical, compile the function (see \link{funC0})}

\item{modelname}{Character, used if \code{compile = TRUE}, sets a fixed filename for the
C file.}

\item{verbose}{Print compiler output to R command line.}
}
\value{
Object of class \link{obsfn}, i.e.
a function \code{y(out, pars, conditions = NULL)} representing the evaluation of the
observation function.
If \code{out} has the attribute  "sensitivities", the result of
\code{y(out, pars)}, will have an attributed "deriv" which reflec the sensitivities of
the observation with respect to the parameters.
If \code{pars} is the result of a parameter transformation \code{p(pars)} (see \link{P}),
the Jacobian
of the parameter transformation and the sensitivities of the observation function
are multiplied according to the chain rule for differentiation.
The original argument \code{out} will be attached to the evaluated observations.
}
\description{
Creates a function \code{y(out, pars)} that evaluates an observation function
and its derivatives based on the output of a model function \code{x(times, pars)}, see \link{Xf} and \link{Xs}.
}

